bdk_chain::bitcoin::base64::engine
Trait Config

pub trait Config {
    // Required method
    fn encode_padding(&self) -> bool;
}

The minimal level of configuration that engines must support.

Required Methods

fn encode_padding(&self) -> bool

Returns true if padding should be added after the encoded output.

Padding is added outside the engine’s encode() since the engine may be used to encode only a chunk of the overall output, so it can’t always know when the output is “done” and would therefore need padding (if configured).