bdk_chain::bitcoin::bech32
Function encode_lower_to_writer

pub fn encode_lower_to_writer<Ck, W>(
    w: &mut W,
    hrp: Hrp,
    data: &[u8],
) -> Result<(), EncodeIoError>
where
    Ck: Checksum,
    W: Write,

Encodes data to a writer (io::Write) as a lowercase bech32 encoded string.

Encoded string will be prefixed with the hrp and have a checksum appended as specified by the Ck algorithm (NoChecksum to exclude checksum all together).