bdk_chain::bitcoin::bech32
Function encode_upper

pub fn encode_upper<Ck>(hrp: Hrp, data: &[u8]) -> Result<String, EncodeError>
where
    Ck: Checksum,

Encodes data as an uppercase bech32 encoded string.

Encoded string will be prefixed with the hrp and have a checksum appended as specified by the Ck algorithm (NoChecksum to exclude checksum all together).