Module bdk_chain::bitcoin::bech32::hrp

Re-exports the hrp types from primitives::hrp to make importing ergonomic for the top level APIs.

Structs

Hrp

The human-readable part (human readable prefix before the ‘1’ separator).

Constants

BC

The human-readable part used by the Bitcoin mainnet network.

BCRT

The human-readable part used when running a Bitcoin regtest network.

TB

The human-readable part used by the Bitcoin testnet networks (testnet, signet).