Module bdk_chain::bitcoin::bech32::primitives

Provides the internal nuts and bolts that enable bech32 encoding/decoding.

Modules

checksum

Degree-2 BCH code checksum.

decode

Decoding of bech32 encoded strings as specified by BIP-173 and BIP-350.

encode

Bech32 address encoding.

gf32

GF32 - Galois Field over 32 elements.

hrp

Provides an Hrp type that represents the human-readable part of a bech32 encoded string.

iter

Iterator Adaptors.

segwit

Segregated Witness functionality - useful for enforcing parts of BIP-173 and BIP-350.

Enums

Bech32

The bech32 checksum algorithm, defined in BIP-173.

Bech32m

The bech32m checksum algorithm, defined in BIP-350.

NoChecksum

The “null checksum” used on bech32 strings for which we want to do no checksum checking.