Struct bdk_chain::bitcoin::bech32::primitives::gf32::Fe32

pub struct Fe32(/* private fields */);

An element in GF(32), the finite field containing elements [0,31] inclusive.

Implementations

impl Fe32

pub const Q: Fe32 = _

Numeric value maps to bech32 character: 0 == “q”.

pub const P: Fe32 = _

Numeric value maps to bech32 character: 1 == “p”.

pub const Z: Fe32 = _

Numeric value maps to bech32 character: 2 == “z”.

pub const R: Fe32 = _

Numeric value maps to bech32 character: 3 == “r”.

pub const Y: Fe32 = _

Numeric value maps to bech32 character: 4 == “y”.

pub const _9: Fe32 = _

Numeric value maps to bech32 character: 5 == “9”.

pub const X: Fe32 = _

Numeric value maps to bech32 character: 6 == “x”.

pub const _8: Fe32 = _

Numeric value maps to bech32 character: 7 == “8”.

pub const G: Fe32 = _

Numeric value maps to bech32 character: 8 == “g”.

pub const F: Fe32 = _

Numeric value maps to bech32 character: 9 == “f”.

pub const _2: Fe32 = _

Numeric value maps to bech32 character: 10 == “2”.

pub const T: Fe32 = _

Numeric value maps to bech32 character: 11 == “t”.

pub const V: Fe32 = _

Numeric value maps to bech32 character: 12 == “v”.

pub const D: Fe32 = _

Numeric value maps to bech32 character: 13 == “d”.

pub const W: Fe32 = _

Numeric value maps to bech32 character: 14 == “w”.

pub const _0: Fe32 = _

Numeric value maps to bech32 character: 15 == “0”.

pub const S: Fe32 = _

Numeric value maps to bech32 character: 16 == “s”.

pub const _3: Fe32 = _

Numeric value maps to bech32 character: 17 == “3”.

pub const J: Fe32 = _

Numeric value maps to bech32 character: 18 == “j”.

pub const N: Fe32 = _

Numeric value maps to bech32 character: 19 == “n”.

pub const _5: Fe32 = _

Numeric value maps to bech32 character: 20 == “5”.

pub const _4: Fe32 = _

Numeric value maps to bech32 character: 21 == “4”.

pub const K: Fe32 = _

Numeric value maps to bech32 character: 22 == “k”.

pub const H: Fe32 = _

Numeric value maps to bech32 character: 23 == “h”.

pub const C: Fe32 = _

Numeric value maps to bech32 character: 24 == “c”.

pub const E: Fe32 = _

Numeric value maps to bech32 character: 25 == “e”.

pub const _6: Fe32 = _

Numeric value maps to bech32 character: 26 == “6”.

pub const M: Fe32 = _

Numeric value maps to bech32 character: 27 == “m”.

pub const U: Fe32 = _

Numeric value maps to bech32 character: 28 == “u”.

pub const A: Fe32 = _

Numeric value maps to bech32 character: 29 == “a”.

pub const _7: Fe32 = _

Numeric value maps to bech32 character: 30 == “7”.

pub const L: Fe32 = _

Numeric value maps to bech32 character: 31 == “l”.

pub fn iter_alpha() -> impl Iterator<Item = Fe32>

Iterator over all field elements, in alphabetical order.

pub fn from_char(c: char) -> Result<Fe32, FromCharError>

Creates a field element from a single bech32 character.

Errors

If the input char is not part of the bech32 alphabet.

pub fn from_char_unchecked(c: u8) -> Fe32

Creates a field element from a single bech32 character.

Panics

If the input character is not part of the bech32 alphabet.

pub fn to_char(self) -> char

Converts the field element to a lowercase bech32 character.

pub fn to_u8(self) -> u8

Converts the field element to a 5-bit u8, with bits representing the coefficients of the polynomial representation.