Module bdk_chain::bitcoin::bech32::primitives::hrp

Provides an Hrp type that represents the human-readable part of a bech32 encoded string.

The human-readable part, which is intended to convey the type of data, or anything else that is relevant to the reader. This part MUST contain 1 to 83 US-ASCII characters, with each character having a value in the range [33-126]. HRP validity may be further restricted by specific applications.

ref: BIP-173

Structs

ByteIter

Iterator over bytes (ASCII values) of the human-readable part.

CharIter

Iterator over ASCII characters of the human-readable part.

Hrp

The human-readable part (human readable prefix before the ‘1’ separator).

LowercaseByteIter

Iterator over lowercase bytes (ASCII characters) of the human-readable part.

LowercaseCharIter

Iterator over lowercase ASCII characters of the human-readable part.

Enums

Error

Errors encountered while checking the human-readable part as defined by BIP-173.

Constants

BC

The human-readable part used by the Bitcoin mainnet network.

BCRT

The human-readable part used when running a Bitcoin regtest network.

TB

The human-readable part used by the Bitcoin testnet networks (testnet, signet).