bdk_chain::bitcoin::bech32
Trait Fe32IterExt

pub trait Fe32IterExt: Sized + Iterator<Item = Fe32> {
    // Provided methods
    fn fes_to_bytes(self) -> FesToBytes<Self>  { ... }
    fn with_checksum<Ck>(self, hrp: &Hrp) -> Encoder<'_, Self, Ck>
       where Ck: Checksum { ... }
}

Extension trait for field element iterators.

Provided Methods

fn fes_to_bytes(self) -> FesToBytes<Self>

Adapts the Fe32 iterator to output bytes instead.

If the total number of bits is not a multiple of 8, any trailing bits are simply dropped.

fn with_checksum<Ck>(self, hrp: &Hrp) -> Encoder<'_, Self, Ck>
where Ck: Checksum,

Adapts the Fe32 iterator to encode the field elements into a bech32 address.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.