bdk_chain::bitcoin
Module bip32

BIP32 implementation.

Implementation of BIP32 hierarchical deterministic wallets, as defined at https://github.com/bitcoin/bips/blob/master/bip-0032.mediawiki.

Structs

ChainCode

A chain code

DerivationPath

A BIP-32 derivation path.

DerivationPathIterator

An iterator over children of a DerivationPath.

Fingerprint

A fingerprint

InvalidBase58PayloadLengthError

Decoded base58 data was an invalid length.

XKeyIdentifier

Extended key identifier as defined in BIP-32.

Xpriv

Extended private key

Xpub

Extended public key

Enums

ChildNumber

A child number for a derived key

Error

A BIP32 error

Traits

IntoDerivationPath

Trait that allows possibly failable conversion from a type into a derivation path

Type Aliases

ExtendedPrivKeyDeprecated

The old name for xpriv, extended public key.

ExtendedPubKeyDeprecated

The old name for xpub, extended public key.

ExtendendPrivKeyDeprecated

The old name for xpriv, extended public key (with a released typo in it).

ExtendendPubKeyDeprecated

The old name for xpub, extended public key (with a released typo in it).

KeySource

Full information on the used extended public key: fingerprint of the master extended public key and a derivation path from it.