Type Alias bdk_chain::bitcoin::bip32::ExtendendPubKey

pub type ExtendendPubKey = Xpub;

๐Deprecated since 0.31.0: use xpub instead

The old name for xpub, extended public key (with a released typo in it).

Aliased Type

struct ExtendendPubKey {
    pub network: NetworkKind,
    pub depth: u8,
    pub parent_fingerprint: Fingerprint,
    pub child_number: ChildNumber,
    pub public_key: PublicKey,
    pub chain_code: ChainCode,
}

Fields
network: NetworkKind

The network kind this key is to be used on

depth: u8

How many derivations this key is from the master (which is 0)

parent_fingerprint: Fingerprint

Fingerprint of the parent key

child_number: ChildNumber

Child number of the key used to derive from parent (0 for master)

public_key: PublicKey

Public key

chain_code: ChainCode

Chain code