Module bdk_chain::bitcoin::blockdata

Bitcoin block data.

This module defines structures and functions for storing the blocks and transactions which make up the Bitcoin system.

Modules

block

Bitcoin blocks.

constants

Blockdata constants.

fee_rate

Implements FeeRate and assoctiated features.

locktime

Provides absolute and relative locktimes.

opcodes

Bitcoin script opcodes.

script

Bitcoin scripts.

transaction

Bitcoin transactions.

weight

Implements Weight and associated features.

witness

Witness

Structs

FeeRate

Represents fee rate.

Weight

Represents block weight - the weight of a transaction or block.