Module bdk_chain::bitcoin::blockdata::block

Bitcoin blocks.

A block is a bundle of transactions with a proof-of-work attached, which commits to an earlier block to form the blockchain. This module describes structures and functions needed to describe these blocks and the blockchain.

Structs

Block

Bitcoin block.

BlockHash

A bitcoin block hash.

Header

Bitcoin block header.

TxMerkleNode

A hash of the Merkle tree branch or root for transactions.

Version

Bitcoin block version number.

WitnessCommitment

A hash corresponding to the witness structure commitment in the coinbase transaction.

WitnessMerkleNode

A hash corresponding to the Merkle tree root for witness data.

Enums

Bip34Error

An error when looking up a BIP34 block height.

ValidationError

A block validation error.