Struct bdk_chain::bitcoin::blockdata::locktime::relative::Time

pub struct Time(/* private fields */);

A relative lock time lock-by-blocktime value.

For BIP 68 relative lock-by-blocktime locks, time is measure in 512 second intervals.

Implementations

impl Time

pub const ZERO: Time = _

Relative block time 0, can be included in any block.

pub const MIN: Time = Time::ZERO

The minimum relative block time (0), can be included in any block.

pub const MAX: Time = _

The maximum relative block time (33,554,432 seconds or approx 388 days).

pub const fn from_512_second_intervals(intervals: u16) -> Time

Create a Time using time intervals where each interval is equivalent to 512 seconds.

Encoding finer granularity of time for relative lock-times is not supported in Bitcoin.

pub const fn from_seconds_floor(seconds: u32) -> Result<Time, TimeOverflowError>

Create a Time from seconds, converting the seconds into 512 second interval with truncating division.

Errors

Will return an error if the input cannot be encoded in 16 bits.

pub const fn from_seconds_ceil(seconds: u32) -> Result<Time, TimeOverflowError>

Create a Time from seconds, converting the seconds into 512 second interval with ceiling division.

Errors

Will return an error if the input cannot be encoded in 16 bits.

pub fn value(self) -> u16

Returns the inner u16 value.

pub fn to_consensus_u32(&self) -> u32

Returns the u32 value used to encode this locktime in an nSequence field or argument to OP_CHECKSEQUENCEVERIFY.