bdk_chain::bitcoin::blockdata::script
Struct Script

pub struct Script(/* private fields */);

Bitcoin script slice.

See also the bitcoin::blockdata::script module.

Script is a script slice, the most primitive script type. It’s usually seen in its borrowed form &Script. It is always encoded as a series of bytes representing the opcodes and data pushes.

Validity

Script does not have any validity invariants - it’s essentially just a marked slice of bytes. This is similar to Path vs OsStr where they are trivially cast-able to each-other and Path doesn’t guarantee being a usable FS path but having a newtype still has value because of added methods, readability and basic type checking.

Although at least data pushes could be checked not to overflow the script, bad scripts are allowed to be in a transaction (outputs just become unspendable) and there even are such transactions in the chain. Thus we must allow such scripts to be placed in the transaction.

Slicing safety

Slicing is similar to how str works: some ranges may be incorrect and indexing by usize is not supported. However, as opposed to std, we have no way of checking correctness without causing linear complexity so there are no panics on invalid ranges! If you supply an invalid range, you’ll get a garbled script.

The range is considered valid if it’s at a boundary of instruction. Care must be taken especially with push operations because you could get a reference to arbitrary attacker-supplied bytes that look like a valid script.

It is recommended to use .instructions() method to get an iterator over script instructions and work with that instead.

Memory safety

The type is #[repr(transparent)] for internal purposes only! No consumer crate may rely on the representation of the struct!

References

Bitcoin Core References

CScript definition

Implementations

impl Script

pub fn new() -> &'static Script

Creates a new empty script.

pub fn from_bytes(bytes: &[u8]) -> &Script

Treat byte slice as Script

pub fn from_bytes_mut(bytes: &mut [u8]) -> &mut Script

Treat mutable byte slice as Script

pub fn as_bytes(&self) -> &[u8]

Returns the script data as a byte slice.

pub fn as_mut_bytes(&mut self) -> &mut [u8]

Returns the script data as a mutable byte slice.

pub fn builder() -> Builder

Creates a new script builder

pub fn script_hash(&self) -> ScriptHash

Returns 160-bit hash of the script.

pub fn wscript_hash(&self) -> WScriptHash

Returns 256-bit hash of the script for P2WSH outputs.

pub fn tapscript_leaf_hash(&self) -> TapLeafHash

Computes leaf hash of tapscript.

pub fn len(&self) -> usize

Returns the length in bytes of the script.

pub fn is_empty(&self) -> bool

Returns whether the script is the empty script.

pub fn to_bytes(&self) -> Vec<u8>

Returns a copy of the script data.

pub fn bytes(&self) -> Bytes<'_>

Returns an iterator over script bytes.

pub fn to_p2wsh(&self) -> ScriptBuf

Computes the P2WSH output corresponding to this witnessScript (aka the “witness redeem script”).

pub fn to_p2tr<C>( &self, secp: &Secp256k1<C>, internal_key: XOnlyPublicKey, ) -> ScriptBuf
where C: Verification,

Computes P2TR output with a given internal key and a single script spending path equal to the current script, assuming that the script is a Tapscript.

pub fn witness_version(&self) -> Option<WitnessVersion>

Returns witness version of the script, if any, assuming the script is a scriptPubkey.

Returns

The witness version if this script is found to conform to the SegWit rules:

A scriptPubKey (or redeemScript as defined in BIP16/P2SH) that consists of a 1-byte push opcode (for 0 to 16) followed by a data push between 2 and 40 bytes gets a new special meaning. The value of the first push is called the “version byte”. The following byte vector pushed is called the “witness program”.

pub fn is_p2sh(&self) -> bool

Checks whether a script pubkey is a P2SH output.

pub fn is_p2pkh(&self) -> bool

Checks whether a script pubkey is a P2PKH output.

pub fn is_push_only(&self) -> bool

Checks whether a script is push only.

Note: OP_RESERVED (0x50) and all the OP_PUSHNUM operations are considered push operations.

pub fn is_p2pk(&self) -> bool

Checks whether a script pubkey is a P2PK output.

You can obtain the public key, if its valid, by calling p2pk_public_key()

pub fn p2pk_public_key(&self) -> Option<PublicKey>

Returns the public key if this script is P2PK with a valid public key.

This may return None even when is_p2pk() returns true. This happens when the public key is invalid (e.g. the point not being on the curve). In this situation the script is unspendable.

pub fn is_multisig(&self) -> bool

Checks whether a script pubkey is a bare multisig output.

In a bare multisig pubkey script the keys are not hashed, the script is of the form:

2 <pubkey1> <pubkey2> <pubkey3> 3 OP_CHECKMULTISIG

pub fn is_witness_program(&self) -> bool

Checks whether a script pubkey is a Segregated Witness (segwit) program.

pub fn is_p2wsh(&self) -> bool

Checks whether a script pubkey is a P2WSH output.

pub fn is_p2wpkh(&self) -> bool

Checks whether a script pubkey is a P2WPKH output.

pub fn is_p2tr(&self) -> bool

Checks whether a script pubkey is a P2TR output.

pub fn is_op_return(&self) -> bool

Check if this is an OP_RETURN output.

pub fn is_provably_unspendable(&self) -> bool

👎Deprecated since 0.32.0: The method has potentially confusing semantics and is going to be removed, you might want is_op_return

Checks whether a script is trivially known to have no satisfying input.

This method has potentially confusing semantics and an unclear purpose, so it’s going to be removed. Use is_op_return if you want OP_RETURN semantics.

pub fn to_p2sh(&self) -> ScriptBuf

Computes the P2SH output corresponding to this redeem script.

pub fn p2wpkh_script_code(&self) -> Option<ScriptBuf>

Returns the script code used for spending a P2WPKH output if this script is a script pubkey for a P2WPKH output. The scriptCode is described in BIP143.

pub fn redeem_script(&self) -> Option<&Script>

Get redeemScript following BIP16 rules regarding P2SH spending.

This does not guarantee that this represents a P2SH input Script. It merely gets the last push of the script. Use Script::is_p2sh on the scriptPubKey to check whether it is actually a P2SH script.

pub fn dust_value(&self) -> Amount

👎Deprecated since 0.32.0: use minimal_non_dust and friends

Returns the minimum value an output with this script should have in order to be broadcastable on today’s Bitcoin network.

pub fn minimal_non_dust(&self) -> Amount

Returns the minimum value an output with this script should have in order to be broadcastable on today’s Bitcoin network.

Dust depends on the -dustrelayfee value of the Bitcoin Core node you are broadcasting to. This function uses the default value of 0.00003 BTC/kB (3 sat/vByte).

To use a custom value, use minimal_non_dust_custom.

pub fn minimal_non_dust_custom(&self, dust_relay_fee: FeeRate) -> Amount

Returns the minimum value an output with this script should have in order to be broadcastable on today’s Bitcoin network.

Dust depends on the -dustrelayfee value of the Bitcoin Core node you are broadcasting to. This function lets you set the fee rate used in dust calculation.

The current default value in Bitcoin Core (as of v26) is 3 sat/vByte.

To use the default Bitcoin Core value, use minimal_non_dust.

pub fn count_sigops(&self) -> usize

Counts the sigops for this Script using accurate counting.

In Bitcoin Core, there are two ways to count sigops, “accurate” and “legacy”. This method uses “accurate” counting. This means that OP_CHECKMULTISIG and its verify variant count for N sigops where N is the number of pubkeys used in the multisig. However, it will count for 20 sigops if CHECKMULTISIG is not preceded by an OP_PUSHNUM from 1 - 16 (this would be an invalid script)

Bitcoin Core uses accurate counting for sigops contained within redeemScripts (P2SH) and witnessScripts (P2WSH) only. It uses legacy for sigops in scriptSigs and scriptPubkeys.

(Note: taproot scripts don’t count toward the sigop count of the block, nor do they have CHECKMULTISIG operations. This function does not count OP_CHECKSIGADD, so do not use this to try and estimate if a taproot script goes over the sigop budget.)

pub fn count_sigops_legacy(&self) -> usize

Counts the sigops for this Script using legacy counting.

In Bitcoin Core, there are two ways to count sigops, “accurate” and “legacy”. This method uses “legacy” counting. This means that OP_CHECKMULTISIG and its verify variant count for 20 sigops.

Bitcoin Core uses legacy counting for sigops contained within scriptSigs and scriptPubkeys. It uses accurate for redeemScripts (P2SH) and witnessScripts (P2WSH).

(Note: taproot scripts don’t count toward the sigop count of the block, nor do they have CHECKMULTISIG operations. This function does not count OP_CHECKSIGADD, so do not use this to try and estimate if a taproot script goes over the sigop budget.)

pub fn instructions(&self) -> Instructions<'_>

Iterates over the script instructions.

Each returned item is a nested enum covering opcodes, datapushes and errors. At most one error will be returned and then the iterator will end. To instead iterate over the script as sequence of bytes call the bytes method.

To force minimal pushes, use instructions_minimal.

pub fn instructions_minimal(&self) -> Instructions<'_>

Iterates over the script instructions while enforcing minimal pushes.

This is similar to instructions but an error is returned if a push is not minimal.

pub fn instruction_indices(&self) -> InstructionIndices<'_>

Iterates over the script instructions and their indices.

Unless the script contains an error, the returned item consists of an index pointing to the position in the script where the instruction begins and the decoded instruction - either an opcode or data push.

To force minimal pushes, use Self::instruction_indices_minimal.

pub fn instruction_indices_minimal(&self) -> InstructionIndices<'_>

Iterates over the script instructions and their indices while enforcing minimal pushes.

This is similar to instruction_indices but an error is returned if a push is not minimal.

pub fn fmt_asm(&self, f: &mut dyn Write) -> Result<(), Error>

Writes the human-readable assembly representation of the script to the formatter.

pub fn to_asm_string(&self) -> String

Returns the human-readable assembly representation of the script.

pub fn to_hex_string(&self) -> String

Formats the script as lower-case hex.

This is a more convenient and performant way to write format!("{:x}", script). For better performance you should generally prefer displaying the script but if String is required (this is common in tests) this method can be used.

pub fn first_opcode(&self) -> Option<Opcode>

Returns the first opcode of the script (if there is any).

pub fn into_script_buf(self: Box<Script>) -> ScriptBuf

Converts a Box<Script> into a ScriptBuf without copying or allocating.