bdk_chain::bitcoin::consensus::serde
Trait EncodeBytes

pub trait EncodeBytes {
    // Required methods
    fn encode_chunk<W>(
        &mut self,
        writer: &mut W,
        bytes: &[u8],
    ) -> Result<(), Error>
       where W: Write;
    fn flush<W>(&mut self, writer: &mut W) -> Result<(), Error>
       where W: Write;
}

Transforms given bytes and writes to the writer.

The encoder is allowed to be buffered (and probably should be). The design passing writer each time bypasses the need for GAT.

Required Methods

fn encode_chunk<W>(&mut self, writer: &mut W, bytes: &[u8]) -> Result<(), Error>
where W: Write,

Transform the provided slice and write to the writer.

This is similar to the write_all method on io::Write.

fn flush<W>(&mut self, writer: &mut W) -> Result<(), Error>
where W: Write,

Write data in buffer (if any) to the writer.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.