bdk_chain::bitcoin
Enum EcdsaSighashType

pub enum EcdsaSighashType {
    All = 1,
    None = 2,
    Single = 3,
    AllPlusAnyoneCanPay = 129,
    NonePlusAnyoneCanPay = 130,
    SinglePlusAnyoneCanPay = 131,
}

Hashtype of an inputโs signature, encoded in the last byte of the signature.

Fixed values so they can be cast as integer types for encoding (see also TapSighashType).

Variants

All = 1

0x1: Sign all outputs.

None = 2

0x2: Sign no outputs โ anyone can choose the destination.

Single = 3

0x3: Sign the output whose index matches this inputโs index. If none exists, sign the hash 0000000000000000000000000000000000000000000000000000000000000001. (This rule is probably an unintentional C++ism, but itโs consensus so we have to follow it.)

AllPlusAnyoneCanPay = 129

0x81: Sign all outputs but only this input.

NonePlusAnyoneCanPay = 130

0x82: Sign no outputs and only this input.

SinglePlusAnyoneCanPay = 131

0x83: Sign one output and only this input (see Single for what โone outputโ means).

Implementations

impl EcdsaSighashType

pub fn from_consensus(n: u32) -> EcdsaSighashType

Creates a EcdsaSighashType from a raw u32.

Note: this replicates consensus behaviour, for current standardness rules correctness you probably want Self::from_standard.

This might cause unexpected behavior because it does not roundtrip. That is, EcdsaSighashType::from_consensus(n) as u32 != n for non-standard values of n. While verifying signatures, the user should retain the n and use it compute the signature hash message.

pub fn from_standard( n: u32, ) -> Result<EcdsaSighashType, NonStandardSighashTypeError>

Creates a EcdsaSighashType from a raw u32.

Errors

If n is a non-standard sighash value.

pub fn to_u32(self) -> u32

Converts EcdsaSighashType to a u32 sighash flag.

The returned value is guaranteed to be a valid according to standardness rules.