Trait bdk_chain::bitcoin::hashes::sha256t::Tag

pub trait Tag {
    // Required method
    fn engine() -> HashEngine ;
}

Trait representing a tag that can be used as a context for SHA256t hashes.

Required Methods

fn engine() -> HashEngine

Returns a hash engine that is pre-tagged and is ready to be used for the data.

Object Safety

This trait is not object safe.