Enum bdk_chain::bitcoin::io::ErrorKind

pub enum ErrorKind {

    NotFound,
    PermissionDenied,
    ConnectionRefused,
    ConnectionReset,
    ConnectionAborted,
    NotConnected,
    AddrInUse,
    AddrNotAvailable,
    BrokenPipe,
    AlreadyExists,
    WouldBlock,
    InvalidInput,
    InvalidData,
    TimedOut,
    WriteZero,
    Interrupted,
    UnexpectedEof,
    Other,

}

A minimal subset of std::io::ErrorKind which is used for Error. Note that, as with std::io, only Self::Interrupted has defined semantics in this crate, all other variants are provided here only to provide higher-fidelity conversions to and from std::io::Error.

Variants

NotFound

An entity was not found, often a file.

PermissionDenied

The operation lacked the necessary privileges to complete.

ConnectionRefused

The connection was refused by the remote server.

ConnectionReset

The connection was reset by the remote server.

ConnectionAborted

The connection was aborted (terminated) by the remote server.

NotConnected

The network operation failed because it was not connected yet.

AddrInUse

A socket address could not be bound because the address is already in use elsewhere.

AddrNotAvailable

A nonexistent interface was requested or the requested address was not local.

BrokenPipe

The operation failed because a pipe was closed.

AlreadyExists

An entity already exists, often a file.

WouldBlock

The operation needs to block to complete, but the blocking operation was requested to not occur.

InvalidInput

A parameter was incorrect.

InvalidData

Data not valid for the operation were encountered.

TimedOut

The I/O operation’s timeout expired, causing it to be canceled.

WriteZero

An error returned when an operation could not be completed because a call to write returned Ok(0).

Interrupted

This operation was interrupted.

UnexpectedEof

An error returned when an operation could not be completed because an “end of file” was reached prematurely.

Other

A custom error that does not fall under any other I/O error kind