bdk_chain::bitcoin::io
Trait BufRead

pub trait BufRead: Read {
    // Required methods
    fn fill_buf(&mut self) -> Result<&[u8], Error>;
    fn consume(&mut self, amount: usize);
}

A trait describing an input stream that uses an internal buffer when reading.

Required Methods

fn fill_buf(&mut self) -> Result<&[u8], Error>

Returns data read from this reader, filling the internal buffer if needed.

fn consume(&mut self, amount: usize)

Marks the buffered data up to amount as consumed.

Panics

May panic if amount is greater than amount of data read by fill_buf.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.