Trait bdk_chain::bitcoin::io::Write

pub trait Write {
    // Required methods
    fn write(&mut self, buf: &[u8]) -> Result<usize, Error>;
    fn flush(&mut self) -> Result<(), Error>;

    // Provided method
    fn write_all(&mut self, buf: &[u8]) -> Result<(), Error> { ... }
}

A generic trait describing an output stream. See std::io::Write for more info.

Required Methods

fn write(&mut self, buf: &[u8]) -> Result<usize, Error>

Writes buf into this writer, returning how many bytes were written.

fn flush(&mut self) -> Result<(), Error>

Flushes this output stream, ensuring that all intermediately buffered contents reach their destination.

Provided Methods

fn write_all(&mut self, buf: &[u8]) -> Result<(), Error>

Attempts to write an entire buffer into this writer.