bdk_chain::bitcoin::p2p

Module message_network

Bitcoin network-related network messages.

This module defines network messages which describe peers and their capabilities.

Structs

Reject

Reject message might be sent by peers rejecting one of our messages

VersionMessage

Some simple messages The version message

Enums

RejectReason

message rejection reason as a code