Module bdk_chain::bitcoin::pow

Proof-of-work related integer types.

Provides the Work and Target types that are used in proof-of-work calculations. The functions here are designed to be fast, by that we mean it is safe to use them to check headers.

Structs

CompactTarget

Encoding of 256-bit target as 32-bit float.

Target

A 256 bit integer representing target.

Work

A 256 bit integer representing work.