bdk_chain::bitcoin

Module psbt

Partially Signed Bitcoin Transactions.

Implementation of BIP174 Partially Signed Bitcoin Transaction Format as defined at https://github.com/bitcoin/bips/blob/master/bip-0174.mediawiki except we define PSBTs containing non-standard sighash types as invalid.

Modules

raw

Raw PSBT key-value pairs.

serialize

PSBT serialization.

Structs

Input

A key-value map for an input of the corresponding index in the unsigned transaction.

Output

A key-value map for an output of the corresponding index in the unsigned transaction.

Psbt

A Partially Signed Transaction.

PsbtSighashType

A Signature hash type for the corresponding input.

Enums

Error

Ways that a Partially Signed Transaction might fail.

ExtractTxError

This error is returned when extracting a Transaction from a Psbt.

GetKeyError

Errors when getting a key.

IndexOutOfBoundsError

Input index out of bounds (actual index, maximum index allowed).

KeyRequest

Data required to call GetKey to get the private key to sign an input.

OutputType

The various output types supported by the Bitcoin network.

PsbtParseError

Error encountered during PSBT decoding from Base64 string.

SignError

Errors encountered while calculating the sighash message.

SigningAlgorithm

Signing algorithms supported by the Bitcoin network.

SigningKeys

A list of keys used to sign an input.

Traits

GetKey

Trait to get a private key from a key request, key is then used to sign an input.

Type Aliases

SigningErrors

Map of input index -> the error encountered while attempting to sign that input.

SigningKeysMap

Map of input index -> signing key for that input (see SigningKeys).