bdk_chain::bitcoin::secp256k1::ffi
Module types

Structs

AlignedType

A type that is as aligned as the biggest alignment for fundamental types in C since C11 that means as aligned as max_align_t is. the exact size/alignment is unspecified.

Enums

c_void

Equivalent to C’s void type when used as a pointer.

Type Aliases

c_char

This might not match C’s c_char exactly. The way we use it makes it fine either way but this type shouldn’t be used outside of the library.

c_int

c_uchar

c_uint

size_t