bdk_chain::bitcoin::secp256k1
Module hashes

👎Deprecated since 0.29.1: Depend on hashes in your own crate.

Deprecated reexport of the bitcoin-hashes crate.

Macros

borrow_slice_impl

Adds slicing traits implementations to a given type $ty

hash_newtype

Creates a new newtype around a Hash type.

hex_fmt_impl

Adds hexadecimal formatting implementation of a trait $imp to a given type $ty.

serde_impl

Implements Serialize and Deserialize for a type $t which represents a newtype over a byte-slice over length $len.

sha256t_hash_newtype

Macro used to define a newtype tagged hash.

Structs

FromSliceError

Attempted to create a hash from an invalid length slice.

Hmac

A hash computed from a RFC 2104 HMAC. Parameterized by the underlying hash function.

HmacEngine

Pair of underlying hash engines, used for the inner and outer hash of HMAC.

Traits

Hash

Trait which applies to hashes of all types.

HashEngine

A hashing engine which bytes can be serialized into.