bdk_chain::bitcoin::secp256k1
Trait PreallocatedContext

pub unsafe trait PreallocatedContext<'a> { }

Trait marking that a particular context object internally points to memory that must outlive 'a

Safety

This trait is used internally to gate which context markers can safely be used with the preallocated_gen_new function. Do not implement it on your own structures.