Struct bdk_chain::bitcoin::Amount

pub struct Amount(/* private fields */);

Amount

The Amount type can be used to express Bitcoin amounts that support arithmetic and conversion to various denominations.

Warning!

This type implements several arithmetic operations from core::ops. To prevent errors due to overflow or underflow when using these operations, it is advised to instead use the checked arithmetic methods whose names start with checked_. The operations from core::ops that Amount implements will panic when overflow or underflow occurs. Also note that since the internal representation of amounts is unsigned, subtracting below zero is considered an underflow and will cause a panic if you’re not using the checked arithmetic methods.

Implementations

impl Amount

pub const ZERO: Amount = _

The zero amount.

pub const ONE_SAT: Amount = _

Exactly one satoshi.

pub const ONE_BTC: Amount = _

Exactly one bitcoin.

pub const MAX_MONEY: Amount = _

The maximum value allowed as an amount. Useful for sanity checking.

pub const MIN: Amount = Amount::ZERO

The minimum value of an amount.

pub const MAX: Amount = _

The maximum value of an amount.

pub const SIZE: usize = 8usize

The number of bytes that an amount contributes to the size of a transaction.

pub const fn from_sat(satoshi: u64) -> Amount

Create an Amount with satoshi precision and the given number of satoshis.

pub fn to_sat(self) -> u64

Gets the number of satoshis in this Amount.

pub fn from_btc(btc: f64) -> Result<Amount, ParseAmountError>

Convert from a value expressing bitcoins to an Amount.

pub const fn from_int_btc(btc: u64) -> Amount

Convert from a value expressing integer values of bitcoins to an Amount in const context.

Panics

The function panics if the argument multiplied by the number of sats per bitcoin overflows a u64 type.

pub fn from_str_in( s: &str, denom: Denomination ) -> Result<Amount, ParseAmountError>

Parse a decimal string as a value in the given denomination.

Note: This only parses the value string. If you want to parse a value with denomination, use FromStr.

pub fn from_str_with_denomination(s: &str) -> Result<Amount, ParseError>

Parses amounts with denomination suffix like they are produced with Self::to_string_with_denomination or with fmt::Display. If you want to parse only the amount without the denomination, use Self::from_str_in.

pub fn to_float_in(self, denom: Denomination) -> f64

Express this Amount as a floating-point value in the given denomination.

Please be aware of the risk of using floating-point numbers.

pub fn to_btc(self) -> f64

Express this Amount as a floating-point value in Bitcoin.

Please be aware of the risk of using floating-point numbers.

Examples

let amount = Amount::from_sat(100_000);
assert_eq!(amount.to_btc(), amount.to_float_in(Denomination::Bitcoin))

pub fn from_float_in( value: f64, denom: Denomination ) -> Result<Amount, ParseAmountError>

Convert this Amount in floating-point notation with a given denomination. Can return error if the amount is too big, too precise or negative.

Please be aware of the risk of using floating-point numbers.

pub fn display_in(self, denomination: Denomination) -> Display

Create an object that implements fmt::Display using specified denomination.

pub fn display_dynamic(self) -> Display

Create an object that implements fmt::Display dynamically selecting denomination.

This will use BTC for values greater than or equal to 1 BTC and satoshis otherwise. To avoid confusion the denomination is always shown.

pub fn fmt_value_in( self, f: &mut dyn Write, denom: Denomination ) -> Result<(), Error>

Format the value of this Amount in the given denomination.

Does not include the denomination.

pub fn to_string_in(self, denom: Denomination) -> String

Get a string number of this Amount in the given denomination.

Does not include the denomination.

pub fn to_string_with_denomination(self, denom: Denomination) -> String

Get a formatted string of this Amount in the given denomination, suffixed with the abbreviation for the denomination.

pub fn checked_add(self, rhs: Amount) -> Option<Amount>

Checked addition.

Returns None if overflow occurred.

pub fn checked_sub(self, rhs: Amount) -> Option<Amount>

Checked subtraction.

Returns None if overflow occurred.

pub fn checked_mul(self, rhs: u64) -> Option<Amount>

Checked multiplication.

Returns None if overflow occurred.

pub fn checked_div(self, rhs: u64) -> Option<Amount>

Checked integer division.

Be aware that integer division loses the remainder if no exact division can be made. Returns None if overflow occurred.

pub fn checked_rem(self, rhs: u64) -> Option<Amount>

Checked remainder.

Returns None if overflow occurred.

pub fn unchecked_add(self, rhs: Amount) -> Amount

Unchecked addition.

Computes self + rhs. Panics in debug mode, wraps in release mode.

pub fn unchecked_sub(self, rhs: Amount) -> Amount

Unchecked subtraction.

Computes self - rhs. Panics in debug mode, wraps in release mode.

pub fn to_signed(self) -> Result<SignedAmount, OutOfRangeError>

Convert to a signed amount.