Struct bdk_chain::bitcoin::taproot::ControlBlock

pub struct ControlBlock {
    pub leaf_version: LeafVersion,
    pub output_key_parity: Parity,
    pub internal_key: XOnlyPublicKey,
    pub merkle_branch: TaprootMerkleBranch,
}

Control block data structure used in Tapscript satisfaction.

Fields
leaf_version: LeafVersion

The tapleaf version.

output_key_parity: Parity

The parity of the output key (NOT THE INTERNAL KEY WHICH IS ALWAYS XONLY).

internal_key: XOnlyPublicKey

The internal key.

merkle_branch: TaprootMerkleBranch

The merkle proof of a script associated with this leaf.

Implementations

impl ControlBlock

pub fn decode(sl: &[u8]) -> Result<ControlBlock, TaprootError>

Decodes bytes representing a ControlBlock.

This is an extra witness element that provides the proof that taproot script pubkey is correctly computed with some specified leaf hash. This is the last element in taproot witness when spending a output via script path.

Errors

TaprootError::InvalidControlBlockSize if sl is not of size 1 + 32 + 32N for any N >= 0.

TaprootError::InvalidTaprootLeafVersion if first byte of sl is not a valid leaf version.

TaprootError::InvalidInternalKey if internal key is invalid (first 32 bytes after the parity byte).

TaprootError::InvalidMerkleTreeDepth if merkle tree is too deep (more than 128 levels).

pub fn size(&self) -> usize

Returns the size of control block. Faster and more efficient than calling Self::serialize().len(). Can be handy for fee estimation.

pub fn encode<W>(&self, writer: &mut W) -> Result<usize, Error>
where W: Write + ?Sized,

Serializes to a writer.

Returns

The number of bytes written to the writer.

pub fn serialize(&self) -> Vec<u8>

Serializes the control block.

This would be required when using ControlBlock as a witness element while spending an output via script path. This serialization does not include the crate::VarInt prefix that would be applied when encoding this element as a witness.

pub fn verify_taproot_commitment<C>( &self, secp: &Secp256k1<C>, output_key: XOnlyPublicKey, script: &Script ) -> bool
where C: Verification,

Verifies that a control block is correct proof for a given output key and script.

Only checks that script is contained inside the taptree described by output key. Full verification must also execute the script with witness data.