Struct bdk_chain::bitcoin::taproot::TapTree

pub struct TapTree(/* private fields */);

Taproot Tree representing a complete binary tree without any hidden nodes.

This is in contrast to NodeInfo, which allows hidden nodes. The implementations for Eq, PartialEq and Hash compare the merkle root of the tree

Implementations

impl TapTree

pub fn node_info(&self) -> &NodeInfo

Gets the reference to inner NodeInfo of this tree root.

pub fn into_node_info(self) -> NodeInfo

Gets the inner NodeInfo of this tree root.

pub fn script_leaves(&self) -> ScriptLeaves<'_>

Returns [TapTreeIter<'_>] iterator for a taproot script tree, operating in DFS order over tree ScriptLeafs.

pub fn root_hash(&self) -> TapNodeHash

Returns the root TapNodeHash of this tree.