Struct bdk_chain::chain_graph::ChainGraph

pub struct ChainGraph<P = TxHeight> { /* private fields */ }

A consistent combination of a SparseChain<P> and a TxGraph<T>.

SparseChain only keeps track of transaction ids and their position in the chain, but you often want to store the full transactions as well. Additionally, you want to make sure that everything in the chain is consistent with the full transaction data. ChainGraph enforces these two invariants:

Every transaction that is in the chain is also in the graph (you always have the full transaction).

No transactions in the chain conflict with each other, i.e., they don’t double spend each other or have ancestors that double spend each other.

Note that the ChainGraph guarantees a 1:1 mapping between transactions in the chain and graph but not the other way around. Transactions may fall out of the chain (via re-org or mempool eviction) but will remain in the graph.

Implementations

impl<P> ChainGraph<P>

pub fn chain(&self) -> &SparseChain<P>

Returns a reference to the internal SparseChain.

pub fn graph(&self) -> &TxGraph

Returns a reference to the internal TxGraph.

impl<P> ChainGraph<P>where
P: ChainPosition,

pub fn new(chain: SparseChain<P>, graph: TxGraph) -> Result<Self, NewError<P>>

Create a new chain graph from a chain and a graph.

There are two reasons this can return an Err:

There is a transaction in the chain that does not have its corresponding full transaction in graph.

The chain has two transactions that are allegedly in it, but they conflict in the graph (so could not possibly be in the same chain).

pub fn inflate_update(
&self,
update: SparseChain<P>,
new_txs: impl IntoIterator<Item = Transaction>
) -> Result<ChainGraph<P>, NewError<P>>

Take an update in the form of a SparseChain<P> and attempt to turn it into a chain graph by filling in full transactions from self and from new_txs. This returns a ChainGraph<P, Cow<T>> where the Cow<'a, T> will borrow the transaction if it got it from self.

This is useful when interacting with services like an electrum server which returns a list of txids and heights when calling script_get_history, which can easily be inserted into a SparseChain<TxHeight>. From there, you need to figure out which full transactions you are missing in your chain graph and form new_txs. You then use inflate_update to turn this into an update ChainGraph<P, Cow<Transaction>> and finally use determine_changeset to generate the changeset from it.

pub fn checkpoint_limit(&self) -> Option<usize>

Gets the checkpoint limit.

Refer to SparseChain::checkpoint_limit for more.

pub fn set_checkpoint_limit(&mut self, limit: Option<usize>)

Sets the checkpoint limit.

Refer to SparseChain::set_checkpoint_limit for more.

pub fn invalidate_checkpoints_preview(&self, from_height: u32) -> ChangeSet<P>

Determines the changes required to invalidate checkpoints from_height (inclusive) and above. Displaced transactions will have their positions moved to TxHeight::Unconfirmed.

pub fn invalidate_checkpoints(&mut self, from_height: u32) -> ChangeSet<P>where
ChangeSet<P>: Clone,

Invalidate checkpoints from_height (inclusive) and above. Displaced transactions will be re-positioned to TxHeight::Unconfirmed.

This is equivalent to calling Self::invalidate_checkpoints_preview and Self::apply_changeset in sequence.

pub fn get_tx_in_chain(&self, txid: Txid) -> Option<(&P, &Transaction)>

Get a transaction currently in the underlying SparseChain.

This does not necessarily mean that it is confirmed in the blockchain; it might just be in the unconfirmed transaction list within the SparseChain.

pub fn insert_tx_preview(
&self,
tx: Transaction,
pos: P
) -> Result<ChangeSet<P>, InsertTxError<P>>

Determines the changes required to insert a transaction into the inner ChainGraph and SparseChain at the given position.

If inserting it into the chain position will result in conflicts, the returned ChangeSet should evict conflicting transactions.

pub fn insert_tx(
&mut self,
tx: Transaction,
pos: P
) -> Result<ChangeSet<P>, InsertTxError<P>>

Inserts [Transaction] at the given chain position.

This is equivalent to calling Self::insert_tx_preview and Self::apply_changeset in sequence.

pub fn insert_txout_preview(
&self,
outpoint: OutPoint,
txout: TxOut
) -> ChangeSet<P>

Determines the changes required to insert a [TxOut] into the internal TxGraph.

pub fn insert_txout(&mut self, outpoint: OutPoint, txout: TxOut) -> ChangeSet<P>

Inserts a [TxOut] into the internal TxGraph.

This is equivalent to calling Self::insert_txout_preview and Self::apply_changeset in sequence.

pub fn insert_checkpoint_preview(
&self,
block_id: BlockId
) -> Result<ChangeSet<P>, InsertCheckpointError>

Determines the changes required to insert a block_id (a height and block hash) into the chain.

If a checkpoint with a different hash already exists at that height, this will return an error.

pub fn insert_checkpoint(
&mut self,
block_id: BlockId
) -> Result<ChangeSet<P>, InsertCheckpointError>

Inserts checkpoint into Self.

This is equivalent to calling Self::insert_checkpoint_preview and Self::apply_changeset in sequence.

pub fn determine_changeset(
&self,
update: &ChainGraph<P>
) -> Result<ChangeSet<P>, UpdateError<P>>

Calculates the difference between self and update in the form of a ChangeSet.

pub fn tx_conflicts_in_chain<'a>(
&'a self,
tx: &'a Transaction
) -> impl Iterator<Item = (&'a P, Txid)> + 'a

Given a transaction, return an iterator of txids that conflict with it (spends at least one of the same inputs). This iterator includes all descendants of conflicting transactions.

This method only returns conflicts that exist in the SparseChain as transactions that are not included in SparseChain are already considered as evicted.

pub fn apply_changeset(&mut self, changeset: ChangeSet<P>)

Applies changeset to self.

Warning this method assumes that the changeset is correctly formed. If it is not, the chain graph may behave incorrectly in the future and panic unexpectedly.

pub fn apply_update(
&mut self,
update: ChainGraph<P>
) -> Result<ChangeSet<P>, UpdateError<P>>

Applies the update chain graph. Note this is shorthand for calling Self::determine_changeset() and Self::apply_changeset() in sequence.

pub fn full_txout(&self, outpoint: OutPoint) -> Option<FullTxOut<P>>

Get the full transaction output at an outpoint if it exists in the chain and the graph.

pub fn transactions_in_chain(
&self
) -> impl DoubleEndedIterator<Item = (&P, &Transaction)>

Iterate over the full transactions and their position in the chain ordered by their position in ascending order.

pub fn spent_by(&self, outpoint: OutPoint) -> Option<(&P, Txid)>

Find the transaction in the chain that spends outpoint.

This uses the input/output relationships in the internal graph. Note that the transaction which includes outpoint does not need to be in the graph or the chain for this to return Some(_).

pub fn is_empty(&self) -> bool

Whether the chain graph contains any data whatsoever.