Enum bdk_chain::CanonicalReason

pub enum CanonicalReason<A> {
    Anchor {
        anchor: A,
        descendant: Option<Txid>,
    },
    ObservedIn {
        observed_in: ObservedIn,
        descendant: Option<Txid>,
    },
}

The reason why a transaction is canonical.

Variants

Anchor

This transaction is anchored in the best chain by A, and therefore canonical.

Fields

anchor: A

The anchor that anchored the transaction in the chain.

descendant: Option<Txid>

Whether the anchor is of the transaction’s descendant.

ObservedIn

This transaction does not conflict with any other transaction with a more recent ObservedIn value or one that is anchored in the best chain.

Fields

observed_in: ObservedIn

The ObservedIn value of the transaction.

descendant: Option<Txid>

Whether the ObservedIn value is of the transaction’s descendant.

Implementations

impl<A: Clone> CanonicalReason<A>

pub fn from_anchor(anchor: A) -> Self

Constructs a CanonicalReason from an anchor.

pub fn from_observed_in(observed_in: ObservedIn) -> Self

Constructs a CanonicalReason from an observed_in value.

pub fn to_transitive(&self, descendant: Txid) -> Self

Contruct a new CanonicalReason from the original which is transitive to descendant.

This signals that either the ObservedIn or Anchor value belongs to the transaction’s descendant, but is transitively relevant.

pub fn descendant(&self) -> &Option<Txid>

This signals that either the ObservedIn or Anchor value belongs to the transaction’s descendant.