bdk_chain::indexer
Module keychain_txout

KeychainTxOutIndex controls how script pubkeys are revealed for multiple keychains and indexes TxOuts with them.

Structs

ChangeSet

Represents updates to the derivation index of a KeychainTxOutIndex. It maps each keychain K to a descriptor and its last revealed index.

KeychainTxOutIndex

KeychainTxOutIndex controls how script pubkeys are revealed for multiple keychains, and indexes TxOuts with them.

Enums

InsertDescriptorError

Error returned from KeychainTxOutIndex::insert_descriptor

Constants

DEFAULT_LOOKAHEAD

The default lookahead for a KeychainTxOutIndex

Traits

FullScanRequestBuilderExt

Trait to extend FullScanRequestBuilder.

SyncRequestBuilderExt

Trait to extend SyncRequestBuilder.