bdk_chain::indexer::keychain_txout
Trait SyncRequestBuilderExt

pub trait SyncRequestBuilderExt<K> {
    // Required methods
    fn revealed_spks_from_indexer<R>(
        self,
        indexer: &KeychainTxOutIndex<K>,
        spk_range: R,
    ) -> Self
       where R: RangeBounds<K>;
    fn unused_spks_from_indexer(self, indexer: &KeychainTxOutIndex<K>) -> Self;
}

Trait to extend SyncRequestBuilder.

Required Methods

fn revealed_spks_from_indexer<R>( self, indexer: &KeychainTxOutIndex<K>, spk_range: R, ) -> Self
where R: RangeBounds<K>,

Add Scripts that are revealed by the indexer of the given spk_range that will be synced against.

fn unused_spks_from_indexer(self, indexer: &KeychainTxOutIndex<K>) -> Self

Add Scripts that are revealed by the indexer but currently unused.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.