Module bdk_chain::keychain

Module for keychain related structures.

A keychain here is a set of application-defined indexes for a miniscript descriptor where we can derive script pubkeys at a particular derivation index. The application’s index is simply anything that implements Ord.

KeychainTxOutIndex indexes script pubkeys of keychains and scans in relevant outpoints (that has a txout containing an indexed script pubkey). Internally, this uses SpkTxOutIndex, but also maintains “revealed” and “lookahead” index counts per keychain.

Structs

Balance

Balance, differentiated into various categories.

ChangeSet

Represents updates to the derivation index of a KeychainTxOutIndex. It maps each keychain K to a descriptor and its last revealed index.

KeychainTxOutIndex

KeychainTxOutIndex controls how script pubkeys are revealed for multiple keychains, and indexes [TxOut]s with them.

NoSuchKeychain

The keychain doesn’t exist. Most likley hasn’t been inserted with KeychainTxOutIndex::insert_descriptor.

Enums

InsertDescriptorError

Error returned from KeychainTxOutIndex::insert_descriptor

Constants

DEFAULT_LOOKAHEAD

The default lookahead for a KeychainTxOutIndex

Type Aliases

Indexed

A tuple of keychain index and T representing the indexed value.

KeychainIndexed

A tuple of keychain K, derivation index (u32) and a T associated with them.