Module bdk_chain::keychain::persist

Persistence for changes made to a KeychainTracker.

BDK’s KeychainTracker needs somewhere to persist changes it makes during operation. Operations like giving out a new address are crucial to persist so that next time the application is loaded, it can find transactions related to that address.

Note that the KeychainTracker does not read this persisted data during operation since it always has a copy in memory.

Structs

Persist

Persist wraps a PersistBackend to create a convenient staging area for changes before they are persisted. Not all changes made to the KeychainTracker need to be written to disk right away so you can use Persist::stage to stage it first and then Persist::commit to finally, write it to disk.

Traits

PersistBackend

A persistence backend for Persist.