Struct bdk_chain::keychain::KeychainTracker

pub struct KeychainTracker<K, P> {
    pub txout_index: KeychainTxOutIndex<K>,
    /* private fields */
}

A convenient combination of a KeychainTxOutIndex and a ChainGraph.

The KeychainTracker atomically updates its KeychainTxOutIndex whenever new chain data is incorporated into its internal ChainGraph.

Fields
txout_index: KeychainTxOutIndex<K>

Index between script pubkeys to transaction outputs

Implementations

impl<K, P> KeychainTracker<K, P>where
P: ChainPosition,
K: Ord + Clone + Debug,

pub fn add_keychain(
&mut self,
keychain: K,
descriptor: Descriptor<DescriptorPublicKey>
)

Add a keychain to the tracker’s txout_index with a descriptor to derive addresses. This is just shorthand for calling KeychainTxOutIndex::add_keychain on the internal txout_index.

Adding a keychain means you will be able to derive new script pubkeys under that keychain and the tracker will discover transaction outputs with those script pubkeys.

pub fn keychains(&mut self) -> &BTreeMap<K, Descriptor<DescriptorPublicKey>>

Get the internal map of keychains to their descriptors. This is just shorthand for calling KeychainTxOutIndex::keychains on the internal txout_index.

pub fn checkpoint_limit(&self) -> Option<usize>

Get the checkpoint limit of the internal SparseChain.

Refer to SparseChain::checkpoint_limit for more.

pub fn set_checkpoint_limit(&mut self, limit: Option<usize>)

Set the checkpoint limit of the internal SparseChain.

Refer to SparseChain::set_checkpoint_limit for more.

pub fn determine_changeset(
&self,
scan: &KeychainScan<K, P>
) -> Result<KeychainChangeSet<K, P>, UpdateError<P>>

Determines the resultant KeychainChangeSet if the given KeychainScan is applied.

Internally, we call ChainGraph::determine_changeset and also determine the additions of KeychainTxOutIndex.

pub fn apply_update(
&mut self,
scan: KeychainScan<K, P>
) -> Result<KeychainChangeSet<K, P>, UpdateError<P>>

Directly applies a KeychainScan on KeychainTracker.

This is equivalent to calling determine_changeset and apply_changeset in sequence.

pub fn apply_changeset(&mut self, changeset: KeychainChangeSet<K, P>)

Applies the changes in changeset to KeychainTracker.

Internally, this calls KeychainTxOutIndex::apply_additions and ChainGraph::apply_changeset in sequence.

pub fn full_txouts(
&self
) -> impl Iterator<Item = (&(K, u32), FullTxOut<P>)> + '_

Iterates through FullTxOuts that are considered to exist in our representation of the blockchain/mempool.

In other words, these are txouts of confirmed and in-mempool transactions, based on our view of the blockchain/mempool.

pub fn full_utxos(&self) -> impl Iterator<Item = (&(K, u32), FullTxOut<P>)> + '_

Iterates through FullTxOuts that are unspent outputs.

Refer to full_txouts for more.

pub fn chain_graph(&self) -> &ChainGraph<P>

Returns a reference to the internal ChainGraph.

pub fn graph(&self) -> &TxGraph

Returns a reference to the internal TxGraph (which is part of the ChainGraph).

pub fn chain(&self) -> &SparseChain<P>

Returns a reference to the internal SparseChain (which is part of the ChainGraph).

pub fn insert_checkpoint_preview(
&self,
block_id: BlockId
) -> Result<KeychainChangeSet<K, P>, InsertCheckpointError>

Determines the changes as a result of inserting block_id (a height and block hash) into the tracker.

The caller is responsible for guaranteeing that a block exists at that height. If a checkpoint already exists at that height with a different hash; this will return an error. Otherwise it will return Ok(true) if the checkpoint didn’t already exist or Ok(false) if it did.

Warning: This function modifies the internal state of the tracker. You are responsible for persisting these changes to disk if you need to restore them.

pub fn insert_checkpoint(
&mut self,
block_id: BlockId
) -> Result<KeychainChangeSet<K, P>, InsertCheckpointError>

Directly insert a block_id into the tracker.

This is equivalent of calling insert_checkpoint_preview and apply_changeset in sequence.

pub fn insert_tx_preview(
&self,
tx: Transaction,
pos: P
) -> Result<KeychainChangeSet<K, P>, InsertTxError<P>>

Determines the changes as a result of inserting a transaction into the inner ChainGraph and optionally into the inner chain at position.

Warning: This function modifies the internal state of the chain graph. You are responsible for persisting these changes to disk if you need to restore them.

pub fn insert_tx(
&mut self,
tx: Transaction,
pos: P
) -> Result<KeychainChangeSet<K, P>, InsertTxError<P>>

Directly insert a transaction into the inner ChainGraph and optionally into the inner chain at position.

This is equivalent of calling insert_tx_preview and apply_changeset in sequence.

pub fn balance(&self, should_trust: impl FnMut(&K) -> bool) -> Balance

Returns the balance of the keychain, i.e., the value of unspent transaction outputs tracked.

The caller provides a should_trust predicate which must decide whether the value of unconfirmed outputs on this keychain are guaranteed to be realized or not. For example:

For an internal (change) keychain, should_trust should generally be true since even if you lose an internal output due to eviction, you will always gain back the value from whatever output the unconfirmed transaction was spending (since that output is presumably from your wallet).

For an external keychain, you might want should_trust to return false since someone may cancel (by double spending) a payment made to addresses on that keychain.

When in doubt set should_trust to return false. This doesn’t do anything other than change where the unconfirmed output’s value is accounted for in Balance.

pub fn balance_at(&self, height: u32) -> u64

Returns the balance of all spendable confirmed unspent outputs of this tracker at a particular height.