Struct bdk_chain::local_chain::LocalChain

pub struct LocalChain { /* private fields */ }

This is a local implementation of ChainOracle.

Implementations

impl LocalChain

pub fn genesis_hash(&self) -> BlockHash

Get the genesis hash.

pub fn from_genesis_hash(hash: BlockHash) -> (Self, ChangeSet)

Construct LocalChain from genesis hash.

pub fn from_changeset(changeset: ChangeSet) -> Result<Self, MissingGenesisError>

Construct a LocalChain from an initial changeset.

pub fn from_tip(tip: CheckPoint) -> Result<Self, MissingGenesisError>

Construct a LocalChain from a given checkpoint tip.

pub fn from_blocks(
blocks: BTreeMap<u32, BlockHash>
) -> Result<Self, MissingGenesisError>

Constructs a LocalChain from a BTreeMap of height to [BlockHash].

The BTreeMap enforces the height order. However, the caller must ensure the blocks are all of the same chain.

pub fn tip(&self) -> CheckPoint

Get the highest checkpoint.

pub fn apply_update(
&mut self,
update: Update
) -> Result<ChangeSet, CannotConnectError>

Applies the given update to the chain.

The method returns ChangeSet on success. This represents the applied changes to self.

There must be no ambiguity about which of the existing chain’s blocks are still valid and which are now invalid. That is, the new chain must implicitly connect to a definite block in the existing chain and invalidate the block after it (if it exists) by including a block at the same height but with a different hash to explicitly exclude it as a connection point.

Additionally, an empty chain can be updated with any chain, and a chain with a single block can have it’s block invalidated by an update chain with a block at the same height but different hash.

Errors

An error will occur if the update does not correctly connect with self.

Refer to Update for more about the update struct.

pub fn apply_changeset(
&mut self,
changeset: &ChangeSet
) -> Result<(), MissingGenesisError>

Apply the given changeset.

pub fn insert_block(
&mut self,
block_id: BlockId
) -> Result<ChangeSet, AlterCheckPointError>

Insert a BlockId.

Errors

Replacing the block hash of an existing checkpoint will result in an error.

pub fn initial_changeset(&self) -> ChangeSet

Derives an initial ChangeSet, meaning that it can be applied to an empty chain to recover the current chain.

pub fn iter_checkpoints(&self) -> CheckPointIter

Iterate over checkpoints in descending height order.

pub fn blocks(&self) -> &BTreeMap<u32, BlockHash>

Get a reference to the internal index mapping the height to block hash.