Trait bdk_chain::persist::PersistBackend

pub trait PersistBackend<C> {
    type WriteError: Debug + Display;
    type LoadError: Debug + Display;

    // Required methods
    fn write_changes(&mut self, changeset: &C) -> Result<(), Self::WriteError>;
    fn load_changes(&mut self) -> Result<Option<C>, Self::LoadError>;
}

A persistence backend for writing and loading changesets.

C represents the changeset; a datatype that records changes made to in-memory data structures that are to be persisted, or retrieved from persistence.

Required Associated Types

type WriteError: Debug + Display

The error the backend returns when it fails to write.

type LoadError: Debug + Display

The error the backend returns when it fails to load changesets C.

Required Methods

fn write_changes(&mut self, changeset: &C) -> Result<(), Self::WriteError>

Writes a changeset to the persistence backend.

It is up to the backend what it does with this. It could store every changeset in a list or it inserts the actual changes into a more structured database. All it needs to guarantee is that load_from_persistence restores a keychain tracker to what it should be if all changesets had been applied sequentially.

fn load_changes(&mut self) -> Result<Option<C>, Self::LoadError>

Return the aggregate changeset C from persistence.