Trait bdk_chain::persist::StageExt

pub trait StageExt: Append + Default + Sized {
    // Provided methods
    fn commit_to<B>(
        &mut self,
        backend: &mut B
    ) -> Result<Option<Self>, B::WriteError>
       where B: PersistBackend<Self> { ... }
    fn append_and_commit_to<B>(
        &mut self,
        changeset: Self,
        backend: &mut B
    ) -> Result<Option<Self>, B::WriteError>
       where B: PersistBackend<Self> { ... }
}

Extends a changeset so that it acts as a convenient staging area for any PersistBackend.

Not all changes to the in-memory representation needs to be written to disk right away. Append::append can be used to stage changes first and then StageExt::commit_to can be used to write changes to disk.

Provided Methods

fn commit_to<B>( &mut self, backend: &mut B ) -> Result<Option<Self>, B::WriteError>
where B: PersistBackend<Self>,

Commit the staged changes to the persistence backend.

Changes that are committed (if any) are returned.

Error

Returns a backend-defined error if this fails.

fn append_and_commit_to<B>( &mut self, changeset: Self, backend: &mut B ) -> Result<Option<Self>, B::WriteError>
where B: PersistBackend<Self>,

Stages a new changeset and commits it (alongside any other previously staged changes) to the persistence backend.

Convenience method for calling Append::append and then StageExt::commit_to.

Object Safety

This trait is not object safe.