Module bdk_chain::sparse_chain

Module for structures that maintain sparse (purposely incomplete) snapshots of blockchain data.

SparseChain stores [Txid]s ordered by an index that implements ChainPosition (this represents the transaction’s position in the blockchain; by default, TxHeight is used). SparseChain also contains “checkpoints” which relate block height to block hash. Changes to a SparseChain is reported by returning ChangeSets.

Updating SparseChain

A sparsechain can be thought of as a consistent snapshot of history. A SparseChain can be updated by applying an update SparseChain on top, but only if they “connect” via their checkpoints and don’t result in unexpected movements of transactions.

// create empty sparsechain
let mut chain = SparseChain::<TxHeight>::default();

/* Updating an empty sparsechain will always succeed */

let update = SparseChain::from_checkpoints(vec![
    BlockId {
        height: 1,
        hash: hash_a,
    },
    BlockId {
        height: 2,
        hash: hash_b,
    },
]);
let _ = chain
    .apply_update(update)
    .expect("updating an empty sparsechain will always succeed");

/* To update a non-empty sparsechain, the update must connect */

let update = SparseChain::from_checkpoints(vec![
    BlockId {
        height: 2,
        hash: hash_b,
    },
    BlockId {
        height: 3,
        hash: hash_c,
    },
]);
let _ = chain
    .apply_update(update)
    .expect("we have connected at block height 2, so this must succeed");

Invalid updates

As shown above, sparsechains can be “connected” by comparing their checkpoints. However, there are situations where two sparsechains cannot connect in a way that guarantees consistency.

// our sparsechain has two checkpoints
let chain = SparseChain::<TxHeight>::from_checkpoints(vec![
    BlockId {
        height: 1,
        hash: hash_a,
    },
    BlockId {
        height: 2,
        hash: hash_b,
    },
]);

/* Example of an ambiguous update that does not fully connect */

let ambiguous_update = SparseChain::from_checkpoints(vec![
    // the update sort of "connects" at checkpoint 1, but...
    BlockId {
        height: 1,
        hash: hash_a,
    },
    // we cannot determine whether checkpoint 3 connects with checkpoint 2
    BlockId {
        height: 3,
        hash: hash_c,
    },
]);
let _ = chain
    .determine_changeset(&ambiguous_update)
    .expect_err("cannot apply ambiguous update");

/* Example of an update that completely misses the point */

let disconnected_update = SparseChain::from_checkpoints(vec![
    // the last checkpoint in the chain is 2, so 3 and 4 do not connect
    BlockId {
        height: 3,
        hash: hash_c,
    },
    BlockId {
        height: 4,
        hash: hash_d,
    },
]);
let _ = chain
    .determine_changeset(&disconnected_update)
    .expect_err("cannot apply a totally-disconnected update");

Handling reorgs

Updates can be formed to evict data from the original sparsechain. This is useful for handling blockchain reorgs.

// our chain has a single checkpoint at height 11.
let mut chain = SparseChain::<TxHeight>::from_checkpoints(vec![BlockId {
    height: 11,
    hash: hash_a,
}]);

// we detect a reorg at height 11, and we introduce a new checkpoint at height 12
let update = SparseChain::from_checkpoints(vec![
    BlockId {
        height: 11,
        hash: hash_b,
    },
    BlockId {
        height: 12,
        hash: hash_c,
    },
]);
let _ = chain
    .apply_update(update)
    .expect("we can evict/replace checkpoint 11 since it is the only checkpoint");

// now our `chain` has two checkpoints (11:hash_b & 12:hash_c)
// we detect another reorg, this time at height 12.
let update = SparseChain::from_checkpoints(vec![
    // we connect at checkpoint 11 as this is our "point of agreement".
    BlockId {
        height: 11,
        hash: hash_b,
    },
    BlockId {
        height: 12,
        hash: hash_d,
    },
]);
let _ = chain
    .apply_update(update)
    .expect("we have provided a valid point of agreement, so our reorg update will succeed");

Movement of transactions during update

If the original sparsechain and update sparsechain contain the same transaction at different ChainPositions, the transaction is considered as “moved”. There are various movements of a transaction that are invalid and update will fail.

Valid movements:

When the transaction moved from unconfirmed (in original) to confirmed (in update). In other words, confirming transactions are allowed!

If there has been a reorg at height x, an originally confirmed transaction at height x or above, may move to another height (that is at x or above, including becoming unconfirmed).

Invalid movements:

A confirmed transaction cannot move without a reorg.

Even with a reorg, an originally confirmed transaction cannot be moved below the height of the reorg.

Custom ChainPosition

SparseChain maintains a list of txids ordered by ChainPosition. By default, TxHeight is used; however, additional data can be incorporated into the implementation.

For example, we can have “perfect ordering” of transactions if our positional index is a combination of block height and transaction position in a block.

#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
pub enum TxPosition {
    Confirmed {
        height: u32,   // height of block
        position: u32, // position of transaction in the block
    },
    Unconfirmed,
}

impl Default for TxPosition {
    fn default() -> Self {
        Self::Unconfirmed
    }
}

impl ChainPosition for TxPosition {
    fn height(&self) -> TxHeight {
        match self {
            Self::Confirmed { height, .. } => TxHeight::Confirmed(*height),
            Self::Unconfirmed => TxHeight::Unconfirmed,
        }
    }

    fn max_ord_of_height(height: TxHeight) -> Self {
        match height {
            TxHeight::Confirmed(height) => Self::Confirmed {
                height,
                position: u32::MAX,
            },
            TxHeight::Unconfirmed => Self::Unconfirmed,
        }
    }

    fn min_ord_of_height(height: TxHeight) -> Self {
        match height {
            TxHeight::Confirmed(height) => Self::Confirmed {
                height,
                position: u32::MIN,
            },
            TxHeight::Unconfirmed => Self::Unconfirmed,
        }
    }
}

let mut chain = SparseChain::<TxPosition>::default();
let _ = chain
    .insert_checkpoint(BlockId {
        height: 10,
        hash: hash_a,
    })
    .unwrap();
let _ = chain
    .insert_tx(
        txid_1,
        TxPosition::Confirmed {
            height: 9,
            position: 4321,
        },
    )
    .unwrap();
let _ = chain
    .insert_tx(
        txid_2,
        TxPosition::Confirmed {
            height: 9,
            position: 1234,
        },
    )
    .unwrap();
let _ = chain
    .insert_tx(
        txid_3,
        TxPosition::Confirmed {
            height: 10,
            position: 321,
        },
    )
    .unwrap();

// transactions are ordered correctly
assert_eq!(
    chain.txids().collect::<Vec<_>>(),
    vec![
        &(
            TxPosition::Confirmed {
                height: 9,
                position: 1234
            },
            txid_2
        ),
        &(
            TxPosition::Confirmed {
                height: 9,
                position: 4321
            },
            txid_1
        ),
        &(
            TxPosition::Confirmed {
                height: 10,
                position: 321
            },
            txid_3
        ),
    ],
);

Structs

ChangeSet

The return value of determine_changeset.

SparseChain

This is a non-monotone structure that tracks relevant [Txid]s that are ordered by chain position P.

Enums

InsertCheckpointError

Represents a failure when trying to insert a checkpoint into SparseChain.

InsertTxError

Represents a failure when trying to insert a [Txid] into SparseChain.

UpdateError

Represents an update failure of SparseChain.

Traits

ChainPosition

Represents a position in which transactions are ordered in SparseChain.