Struct bdk_chain::sparse_chain::SparseChain

pub struct SparseChain<P = TxHeight> { /* private fields */ }

This is a non-monotone structure that tracks relevant [Txid]s that are ordered by chain position P.

We use [BlockHash]s alongside their chain height as “checkpoints” to enforce consistency.

To “merge” two SparseChains, the ChangeSet can be calculated by calling determine_changeset and applying the ChangeSet via apply_changeset. For convenience, apply_update does the above two steps in one call.

Refer to module-level documentation for more.

Implementations

impl<P: ChainPosition> SparseChain<P>

pub fn from_checkpoints<C>(checkpoints: C) -> Selfwhere
C: IntoIterator<Item = BlockId>,

Creates a new chain from a list of block hashes and heights. The caller must guarantee they are in the same chain.

pub fn latest_checkpoint(&self) -> Option<BlockId>

Get the checkpoint for the last known tip.

pub fn checkpoint_at(&self, height: u32) -> Option<BlockId>

Get the checkpoint at the given height if it exists.

pub fn tx_position(&self, txid: Txid) -> Option<&P>

Return the ChainPosition of a txid.

This returns None if the transaction does not exist.

pub fn checkpoints(&self) -> &BTreeMap<u32, BlockHash>

Return a BTreeMap of all checkpoints (block hashes by height).

pub fn range_checkpoints(
&self,
range: impl RangeBounds<u32>
) -> impl DoubleEndedIterator<Item = BlockId> + '_

Return an iterator over checkpoints in a height range, in ascending height order.

pub fn determine_changeset(
&self,
update: &Self
) -> Result<ChangeSet<P>, UpdateError<P>>

Preview changes of updating Self with another chain that connects to it.

If the update wishes to introduce confirmed transactions, it must contain a checkpoint that is exactly the same height as one of self’s checkpoints.

To invalidate from a given checkpoint, update must contain a checkpoint of the same height but different hash. Invalidated checkpoints result in invalidated transactions becoming “unconfirmed”.

An error will be returned if an update results in inconsistencies or if the update does not correctly connect with self.

Refer to module-level documentation for more.

pub fn apply_update(
&mut self,
update: Self
) -> Result<ChangeSet<P>, UpdateError<P>>

Updates SparseChain with another chain that connects to it.

This is equivilant to calling determine_changeset and apply_changeset in sequence.

pub fn apply_changeset(&mut self, changeset: ChangeSet<P>)

pub fn initial_changeset(&self) -> ChangeSet<P>

Derives a ChangeSet that assumes that there are no preceding changesets.

The changeset returned will record additions of all [Txid]s and checkpoints included in Self.

pub fn invalidate_checkpoints_preview(&self, from_height: u32) -> ChangeSet<P>

Determines the ChangeSet when checkpoints from_height (inclusive) and above are invalidated. Displaced [Txid]s will be repositioned to TxHeight::Unconfirmed.

pub fn invalidate_checkpoints(&mut self, from_height: u32) -> ChangeSet<P>

Invalidate checkpoints from_height (inclusive) and above.

This is equivalent to calling invalidate_checkpoints_preview and apply_changeset in sequence.

pub fn clear_mempool_preview(&self) -> ChangeSet<P>

Determines the ChangeSet when all transactions of height TxHeight::Unconfirmed are removed completely.

pub fn clear_mempool(&mut self) -> ChangeSet<P>

Clears all transactions of height TxHeight::Unconfirmed.

This is equivalent to calling clear_mempool_preview and apply_changeset in sequence.

ChangeSet.

pub fn insert_tx_preview(
&self,
txid: Txid,
pos: P
) -> Result<ChangeSet<P>, InsertTxError<P>>

Determines the resultant ChangeSet if [Txid] was inserted at position pos.

Changes to the [Txid]’s position are allowed (under the rules noted in module-level documentation) and will be reflected in the ChangeSet.

pub fn insert_tx(
&mut self,
txid: Txid,
pos: P
) -> Result<ChangeSet<P>, InsertTxError<P>>

Inserts a given [Txid] at pos.

This is equivilant to calling insert_tx_preview and apply_changeset in sequence.

pub fn insert_checkpoint_preview(
&self,
block_id: BlockId
) -> Result<ChangeSet<P>, InsertCheckpointError>

Determines the resultant ChangeSet if BlockId was inserted.

If the change would result in a change in block hash of a certain height, insertion would fail.

pub fn insert_checkpoint(
&mut self,
block_id: BlockId
) -> Result<ChangeSet<P>, InsertCheckpointError>

Insert a checkpoint (BlockId).

This is equivalent to calling insert_checkpoint_preview and apply_changeset in sequence.

pub fn txids(
&self
) -> impl DoubleEndedIterator<Item = &(P, Txid)> + ExactSizeIterator + '_

Iterate over all [Txid]s ordered by their ChainPosition.

pub fn range_txids<R>(
&self,
range: R
) -> impl DoubleEndedIterator<Item = &(P, Txid)> + '_where
R: RangeBounds<(P, Txid)>,

Iterate over a sub-range of positioned [Txid]s.

pub fn range_txids_by_position<R>(
&self,
range: R
) -> impl DoubleEndedIterator<Item = &(P, Txid)> + '_where
R: RangeBounds<P>,

Iterate over a sub-range of positioned [Txid]s, where the range is defined by ChainPosition only.

pub fn range_txids_by_height<R>(
&self,
range: R
) -> impl DoubleEndedIterator<Item = &(P, Txid)> + '_where
R: RangeBounds<TxHeight>,

Iterate over a sub-range of positioned [Txid]s, where the range is defined by TxHeight only.

pub fn full_txout(
&self,
graph: &TxGraph,
outpoint: OutPoint
) -> Option<FullTxOut<P>>

Attempt to retrieve a FullTxOut of the given outpoint.

This will return Some only if the output’s transaction is in both self and graph.

pub fn checkpoint_limit(&self) -> Option<usize>

Returns the value set as the checkpoint limit.

Refer to set_checkpoint_limit.

pub fn set_checkpoint_limit(&mut self, limit: Option<usize>)

Set the checkpoint limit.

The checkpoint limit restricts the number of checkpoints that can be stored in Self. Oldest checkpoints are pruned first.

pub fn changeset_additions<'a>(
&'a self,
changeset: &'a ChangeSet<P>
) -> impl Iterator<Item = Txid> + 'a

Return [Txid]s that would be added to the sparse chain if this changeset was applied.

pub fn spent_by(&self, graph: &TxGraph, outpoint: OutPoint) -> Option<(&P, Txid)>

Finds the transaction in the chain that spends outpoint.

TxGraph is used to provide the spend relationships.

Note that the transaction including outpoint does not need to be in the graph or the chain for this to return Some.

pub fn is_empty(&self) -> bool

Returns whether the sparse chain contains any checkpoints or transactions.