Struct bdk_chain::spk_client::FullScanRequest

pub struct FullScanRequest<K> { /* private fields */ }

Data required to perform a spk-based blockchain client full scan.

A client full scan iterates through all the scripts for the given keychains, fetching relevant data until some stop gap number of scripts is found that have no data. This operation is generally only used when importing or restoring previously used keychains in which the list of used scripts is not known. The full scan process also updates the chain from the given chain_tip (if provided).

Implementations

impl<K> FullScanRequest<K>
where K: Ord + Clone,

pub fn builder() -> FullScanRequestBuilder<K>

Start building a FullScanRequest.

pub fn chain_tip(&self) -> Option<CheckPoint>

Get the chain tip CheckPoint of this request (if any).

pub fn keychains(&self) -> Vec<K>

List all keychains contained in this request.

pub fn next_spk(&mut self, keychain: K) -> Option<(u32, ScriptBuf)>

Advances the full scan request and returns the next indexed ScriptBuf of the given keychain.

pub fn iter_spks( &mut self, keychain: K ) -> impl Iterator<Item = (u32, ScriptBuf)>

Iterate over indexed ScriptBufs contained in this request of the given keychain.