Trait bdk_chain::AnchorFromBlockPosition

pub trait AnchorFromBlockPosition: Anchor {
    // Required method
    fn from_block_position(
        block: &Block,
        block_id: BlockId,
        tx_pos: usize
    ) -> Self;
}

An Anchor that can be constructed from a given block, block height and transaction position within the block.

Required Methods

fn from_block_position(block: &Block, block_id: BlockId, tx_pos: usize) -> Self

Construct the anchor from a given block, block height and tx_pos within the block.

Object Safety

This trait is not object safe.