Trait bdk_chain::DescriptorExt

pub trait DescriptorExt {
    // Required methods
    fn dust_value(&self) -> u64;
    fn descriptor_id(&self) -> DescriptorId;
}

A trait to extend the functionality of a miniscript descriptor.

Required Methods

fn dust_value(&self) -> u64

Returns the minimum value (in satoshis) at which an output is broadcastable. Panics if the descriptor wildcard is hardened.

fn descriptor_id(&self) -> DescriptorId

Returns the descriptor ID, calculated as the sha256 hash of the spk derived from the descriptor at index 0.