Trait bdk_chain::ForEachTxOut

pub trait ForEachTxOut {
    fn for_each_txout(&self, f: impl FnMut((OutPoint, &TxOut)));
}

Trait to do something with every txout contained in a structure.

We would prefer to just work with things that can give us an Iterator<Item=(OutPoint, &TxOut)> here, but rust’s type system makes it extremely hard to do this (without trait objects).

Required Methods

fn for_each_txout(&self, f: impl FnMut((OutPoint, &TxOut)))

The provided closure f will be called with each outpoint/txout pair.