Trait bdk_chain::Merge

pub trait Merge: Default {
    // Required methods
    fn merge(&mut self, other: Self);
    fn is_empty(&self) -> bool;

    // Provided method
    fn take(&mut self) -> Option<Self> { ... }
}

Trait that makes an object mergeable.

Required Methods

fn merge(&mut self, other: Self)

Merge another object of the same type onto self.

fn is_empty(&self) -> bool

Returns whether the structure is considered empty.

Provided Methods

fn take(&mut self) -> Option<Self>

Take the value, replacing it with the default value.

Object Safety

This trait is not object safe.