Struct bdk_chain::tx_graph::TxGraph

pub struct TxGraph { /* private fields */ }

A graph of transactions and spends.

See the module-level documentation for more.

Implementations

impl TxGraph

pub fn all_txouts(&self) -> impl Iterator<Item = (OutPoint, &TxOut)>

Iterate over all tx outputs known by TxGraph.

pub fn full_transactions(&self) -> impl Iterator<Item = &Transaction>

Iterate over all full transactions in the graph.

pub fn get_tx(&self, txid: Txid) -> Option<&Transaction>

Get a transaction by txid. This only returns Some for full transactions.

Refer to get_txout for getting a specific [TxOut].

pub fn get_txout(&self, outpoint: OutPoint) -> Option<&TxOut>

Obtains a single tx output (if any) at the specified outpoint.

pub fn txouts(&self, txid: Txid) -> Option<BTreeMap<u32, &TxOut>>

Returns a BTreeMap of vout to output of the provided txid.

pub fn calculate_fee(&self, tx: &Transaction) -> Option<i64>

Calculates the fee of a given transaction. Returns 0 if tx is a coinbase transaction. Returns Some(_) if we have all the TxOuts being spent by tx in the graph (either as the full transactions or individual txouts). If the returned value is negative, then the transaction is invalid according to the graph.

Returns None if we’re missing an input for the tx in the graph.

Note tx does not have to be in the graph for this to work.

impl TxGraph

pub fn new(txs: impl IntoIterator<Item = Transaction>) -> Self

Construct a new TxGraph from a list of transactions.

pub fn insert_txout(&mut self, outpoint: OutPoint, txout: TxOut) -> Additions

Inserts the given [TxOut] at [OutPoint].

Note this will ignore the action if we already have the full transaction that the txout is alleged to be on (even if it doesn’t match it!).

pub fn insert_tx(&mut self, tx: Transaction) -> Additions

Inserts the given transaction into TxGraph.

The Additions returned will be empty if tx already exists.

pub fn apply_update(&mut self, update: TxGraph) -> Additions

Extends this graph with another so that self becomes the union of the two sets of transactions.

The returned Additions is the set difference between update and self (transactions that exist in update but not in self).

pub fn apply_additions(&mut self, additions: Additions)

Applies Additions to TxGraph.

pub fn determine_additions(&self, update: &TxGraph) -> Additions

Previews the resultant Additions when Self is updated against the update graph.

The Additions would be the set difference between update and self (transactions that exist in update but not in self).

pub fn insert_tx_preview(&self, tx: Transaction) -> Additions

Returns the resultant Additions if the given transaction is inserted. Does not actually mutate Self.

The Additions result will be empty if tx already exists in self.

pub fn insert_txout_preview(&self, outpoint: OutPoint, txout: TxOut) -> Additions

Returns the resultant Additions if the given txout is inserted at outpoint. Does not mutate self.

The Additions result will be empty if the outpoint (or a full transaction containing the outpoint) already existed in self.

impl TxGraph

pub fn outspends(&self, outpoint: OutPoint) -> &HashSet<Txid>

The transactions spending from this output.

TxGraph allows conflicting transactions within the graph. Obviously the transactions in the returned set will never be in the same active-chain.

pub fn tx_outspends(
&self,
txid: Txid
) -> impl DoubleEndedIterator<Item = (u32, &HashSet<Txid>)> + '_

Iterates over the transactions spending from txid.

The iterator item is a union of (vout, txid-set) where:

vout is the provided txid’s outpoint that is being spent

txid-set is the set of txids spending the vout.

pub fn partial_transactions(
&self
) -> impl Iterator<Item = (Txid, &BTreeMap<u32, TxOut>)>

Iterate over all partial transactions (outputs only) in the graph.

pub fn walk_descendants<'g, F, O>(
&'g self,
txid: Txid,
walk_map: F
) -> TxDescendants<'_, F> where
F: FnMut(usize, Txid) -> Option<O> + 'g,

Creates an iterator that filters and maps descendants from the starting txid.

The supplied closure takes in two inputs (depth, descendant_txid):

depth is the distance between the starting txid and the descendant_txid. I.e., if the descendant is spending an output of the starting txid; the depth will be 1.

descendant_txid is the descendant’s txid which we are considering to walk.

The supplied closure returns an Option<T>, allowing the caller to map each node it vists and decide whether to visit descendants.

pub fn walk_conflicts<'g, F, O>(
&'g self,
tx: &'g Transaction,
walk_map: F
) -> TxDescendants<'_, F> where
F: FnMut(usize, Txid) -> Option<O> + 'g,

Creates an iterator that both filters and maps conflicting transactions (this includes descendants of directly-conflicting transactions, which are also considered conflicts).

Refer to Self::walk_descendants for walk_map usage.

pub fn direct_conflicts_of_tx<'g>(
&'g self,
tx: &'g Transaction
) -> impl Iterator<Item = (usize, Txid)> + '_

Given a transaction, return an iterator of txids that directly conflict with the given transaction’s inputs (spends). The conflicting txids are returned with the given transaction’s vin (in which it conflicts).

Note that this only returns directly conflicting txids and does not include descendants of those txids (which are technically also conflicting).

pub fn is_empty(&self) -> bool

Whether the graph has any transactions or outputs in it.