Crate bdk_chain

This crate is a collection of core structures for Bitcoin Dev Kit.

The goal of this crate is to give wallets the mechanisms needed to:

Figure out what data they need to fetch.

Process the data in a way that never leads to inconsistent states.

Fully index that data and expose it to be consumed without friction.

Our design goals for these mechanisms are:

Data source agnostic – nothing in bdk_chain cares about where you get data from or whether you do it synchronously or asynchronously. If you know a fact about the blockchain, you can just tell bdk_chain’s APIs about it, and that information will be integrated, if it can be done consistently.

Data persistence agnostic – bdk_chain does not care where you cache on-chain data, what you cache or how you retrieve it from persistent storage.

Re-exports

pub extern crate serde_crate as serde;

pub use bitcoin;

pub use indexed_tx_graph::IndexedTxGraph;

pub use tx_graph::TxGraph;

pub use miniscript;

Modules

indexed_tx_graph

Contains the IndexedTxGraph and associated types. Refer to the IndexedTxGraph documentation for more.

keychain

Module for keychain related structures.

local_chain

The LocalChain is a local implementation of ChainOracle.

tx_graph

Module for structures that store and traverse transactions.

Structs

BlockId

A reference to a block in the canonical chain.

ConfirmationHeightAnchor

An Anchor implementation that also records the exact confirmation height of the transaction.

ConfirmationTimeHeightAnchor

An Anchor implementation that also records the exact confirmation time and height of the transaction.

FullTxOut

A TxOut with as much data as we can retrieve about it

Persist

Persist wraps a PersistBackend (B) to create a convenient staging area for changes (C) before they are persisted.

SpkIterator

An iterator for derived script pubkeys.

SpkTxOutIndex

An index storing TxOuts that have a script pubkey that matches those in a list.

Enums

ChainPosition

Represents the observed position of some chain data.

ConfirmationTime

Block height and timestamp at which a transaction is confirmed.

Constants

BIP32_MAX_INDEX

Maximum BIP32 derivation index.

COINBASE_MATURITY

How many confirmations are needed f or a coinbase output to be spent.

Traits

Anchor

Trait that “anchors” blockchain data to a specific block of height and hash.

AnchorFromBlockPosition

An Anchor that can be constructed from a given block, block height and transaction position within the block.

Append

Trait that makes an object appendable.

ChainOracle

Represents a service that tracks the blockchain.

DescriptorExt

A trait to extend the functionality of a miniscript descriptor.

PersistBackend

A persistence backend for Persist.