bdk_chain::bitcoin::blockdata::locktime

Module absolute

Provides type LockTime that implements the logic around nLockTime/OP_CHECKLOCKTIMEVERIFY.

There are two types of lock time: lock-by-blockheight and lock-by-blocktime, distinguished by whether LockTime < LOCKTIME_THRESHOLD.

Structs

ConversionError

An error that occurs when converting a u32 to a lock time variant.

Height

An absolute block height, guaranteed to always contain a valid height value.

ParseHeightError

Error returned when parsing block height fails.

ParseTimeError

Error returned when parsing block time fails.

Time

A UNIX timestamp, seconds since epoch, guaranteed to always contain a valid time value.

Enums

LockTime

An absolute lock time value, representing either a block height or a UNIX timestamp (seconds since epoch).

Constants

LOCK_TIME_THRESHOLD

The Threshold for deciding whether a lock time value is a height or a time (see Bitcoin Core).