Module bdk_chain::bitcoin::blockdata::locktime::relative

Provides type LockTime that implements the logic around nSequence/OP_CHECKSEQUENCEVERIFY.

There are two types of lock time: lock-by-blockheight and lock-by-blocktime, distinguished by whether bit 22 of the u32 consensus value is set.

Structs

DisabledLockTimeError

Error returned when a sequence number is parsed as a lock time, but its “disable” flag is set.

Height

A relative lock time lock-by-blockheight value.

IncompatibleHeightError

Tried to satisfy a lock-by-blocktime lock using a height value.

IncompatibleTimeError

Tried to satisfy a lock-by-blockheight lock using a time value.

Time

A relative lock time lock-by-blocktime value.

TimeOverflowError

Input time in seconds was too large to be encoded to a 16 bit 512 second interval.

Enums

LockTime

A relative lock time value, representing either a block height or time (512 second intervals).