bdk_chain::bitcoin::hashes::sha256t
Trait Tag

pub trait Tag {
    // Required method
    fn engine() -> HashEngine ;
}

Trait representing a tag that can be used as a context for SHA256t hashes.

Required Methods

fn engine() -> HashEngine

Returns a hash engine that is pre-tagged and is ready to be used for the data.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.