Module bdk_chain::bitcoin::p2p

Bitcoin p2p network types.

This module defines support for (de)serialization and network transport of Bitcoin data and Bitcoin p2p network messages.

Modules

address

Bitcoin network addresses.

message

Bitcoin network messages.

message_blockdata

Bitcoin blockdata network messages.

message_bloom

Bitcoin Connection Bloom filtering network messages.

message_compact_blocks

BIP152 Compact Blocks network messages

message_filter

Bitcoin Client Side Block Filtering network messages.

message_network

Bitcoin network-related network messages.

Structs

Address

A message which can be sent on the Bitcoin network

Magic

Network magic bytes to identify the cryptocurrency network the message was intended for.

ParseMagicError

An error in parsing magic bytes.

ServiceFlags

Flags to indicate which network services a node supports.

UnknownMagicError

Error in creating a Network from Magic bytes.

Constants

PROTOCOL_VERSION

Version of the protocol as appearing in network message headers.