Trait bdk_chain::bitcoin::psbt::GetKey

pub trait GetKey {
    type Error: Debug;

    // Required method
    fn get_key<C>(
        &self,
        key_request: KeyRequest,
        secp: &Secp256k1<C>
    ) -> Result<Option<PrivateKey>, Self::Error>
       where C: Signing;
}

Trait to get a private key from a key request, key is then used to sign an input.

Required Associated Types

type Error: Debug

An error occurred while getting the key.

Required Methods

fn get_key<C>( &self, key_request: KeyRequest, secp: &Secp256k1<C> ) -> Result<Option<PrivateKey>, Self::Error>
where C: Signing,

Attempts to get the private key for key_request.

Returns

Some(key) if the key is found.

None if the key was not found but no error was encountered.

Err if an error was encountered while looking for the key.

Object Safety

This trait is not object safe.