Module bdk_chain::keychain

Module for keychain related structures.

A keychain here is a set of application-defined indexes for a miniscript descriptor where we can derive script pubkeys at a particular derivation index. The application’s index is simply anything that implements Ord.

KeychainTxOutIndex indexes script pubkeys of keychains and scans in relevant outpoints (that has a txout containing an indexed script pubkey). Internally, this uses SpkTxOutIndex, but also maintains “revealed” and “lookahead” index counts per keychain.

KeychainTracker combines ChainGraph and KeychainTxOutIndex and enforces atomic changes between both these structures. KeychainScan is a structure used to update to KeychainTracker and changes made on a KeychainTracker are reported by KeychainChangeSets.

Re-exports

pub use persist::*;

Modules

persist

Persistence for changes made to a KeychainTracker.

Structs

Balance

Balance, differentiated into various categories.

DerivationAdditions

Represents updates to the derivation index of a KeychainTxOutIndex.

KeychainChangeSet

Represents changes to a KeychainTracker.

KeychainScan

An update that includes the last active indexes of each keychain.

KeychainTracker

A convenient combination of a KeychainTxOutIndex and a ChainGraph.

KeychainTxOutIndex

A convenient wrapper around SpkTxOutIndex that relates script pubkeys to miniscript public Descriptors.

Constants

BIP32_MAX_INDEX

Maximum BIP32 derivation index.