Module bdk_chain::local_chain

The LocalChain is a local implementation of ChainOracle.

Structs

AlterCheckPointError

Represents a failure when trying to insert/remove a checkpoint to/from LocalChain.

CannotConnectError

Occurs when an update does not have a common checkpoint with the original chain.

CheckPoint

A LocalChain checkpoint is used to find the agreement point between two chains and as a transaction anchor.

CheckPointIter

Iterates over checkpoints backwards.

LocalChain

This is a local implementation of ChainOracle.

MissingGenesisError

An error which occurs when a LocalChain is constructed without a genesis checkpoint.

Update

Used to update LocalChain.

Enums

ApplyHeaderError

The error type for LocalChain::apply_header_connected_to.

Type Definitions

ChangeSet

The ChangeSet represents changes to LocalChain.