bdk_chain::spk_client
Struct SyncRequest

pub struct SyncRequest<I = ()> { /* private fields */ }

Data required to perform a spk-based blockchain client sync.

A client sync fetches relevant chain data for a known list of scripts, transaction ids and outpoints. The sync process also updates the chain from the given chain_tip (if provided).

// Construct a sync request.
let sync_request = SyncRequest::builder()
    // Provide chain tip of the local wallet.
    .chain_tip(local_chain.tip())
    // Provide list of scripts to scan for transactions against.
    .spks(scripts)
    // This is called for every synced item.
    .inspect(|item, progress| println!("{} (remaining: {})", item, progress.remaining()))
    // Finish constructing the sync request.
    .build();

Implementations

impl<I> SyncRequest<I>

pub fn builder() -> SyncRequestBuilder<I>

Start building a SyncRequest.

pub fn progress(&self) -> SyncProgress

Get the SyncProgress of this request.

pub fn chain_tip(&self) -> Option<CheckPoint>

Get the chain tip CheckPoint of this request (if any).

pub fn next_spk(&mut self) -> Option<ScriptBuf>

Advances the sync request and returns the next ScriptBuf.

Returns None when there are no more scripts remaining in the request.

pub fn next_txid(&mut self) -> Option<Txid>

Advances the sync request and returns the next Txid.

Returns None when there are no more txids remaining in the request.

pub fn next_outpoint(&mut self) -> Option<OutPoint>

Advances the sync request and returns the next OutPoint.

Returns None when there are no more outpoints in the request.

pub fn iter_spks(&mut self) -> impl ExactSizeIterator

Iterate over ScriptBufs contained in this request.

pub fn iter_txids(&mut self) -> impl ExactSizeIterator

Iterate over Txids contained in this request.

pub fn iter_outpoints(&mut self) -> impl ExactSizeIterator

Iterate over OutPoints contained in this request.