bdk_chain
Trait DescriptorExt

pub trait DescriptorExt {
    // Required methods
    fn dust_value(&self) -> Amount;
    fn descriptor_id(&self) -> DescriptorId;
}

A trait to extend the functionality of a miniscript descriptor.

Required Methods

fn dust_value(&self) -> Amount

Returns the minimum Amount at which an output is broadcast-able. Panics if the descriptor wildcard is hardened.

fn descriptor_id(&self) -> DescriptorId

Returns the descriptor ID, calculated as the sha256 hash of the spk derived from the descriptor at index 0.