Module bdk_chain::tx_graph

Module for structures that store and traverse transactions.

TxGraph contains transactions and indexes them so you can easily traverse the graph of those transactions. TxGraph is monotone in that you can always insert a transaction – it doesn’t care whether that transaction is in the current best chain or whether it conflicts with any of the existing transactions or what order you insert the transactions. This means that you can always combine two TxGraphs together, without resulting in inconsistencies. Furthermore, there is currently no way to delete a transaction.

Transactions can be either whole or partial (i.e., transactions for which we only know some outputs, which we usually call “floating outputs”; these are usually inserted using the insert_txout method.).

The graph contains transactions in the form of TxNodes. Each node contains the txid, the transaction (whole or partial), the blocks it’s anchored in (see the Anchor documentation for more details), and the timestamp of the last time we saw the transaction as unconfirmed.

Conflicting transactions are allowed to coexist within a TxGraph. This is useful for identifying and traversing conflicts and descendants of a given transaction. Some TxGraph methods only consider “canonical” (i.e., in the best chain or in mempool) transactions, we decide which transactions are canonical based on anchors last_seen_unconfirmed; see the try_get_chain_position documentation for more details.

The ChangeSet reports changes made to a TxGraph; it can be used to either save to persistent storage, or to be applied to another TxGraph.

Lastly, you can use TxAncestors/TxDescendants to traverse ancestors and descendants of a given transaction, respectively.

Applying changes

Methods that apply changes to TxGraph will return ChangeSet. ChangeSet can be applied back to a TxGraph or be used to inform persistent storage of the changes to TxGraph.

let mut graph: TxGraph = TxGraph::default();
let mut another_graph: TxGraph = TxGraph::default();

// insert a transaction
let changeset = graph.insert_tx(tx_a);

// the resulting changeset can be applied to another tx graph
another_graph.apply_changeset(changeset);

A TxGraph can also be updated with another TxGraph.

let mut graph: TxGraph = TxGraph::default();
let update = TxGraph::new(vec![tx_a, tx_b]);

// apply the update graph
let changeset = graph.apply_update(update.clone());

// if we apply it again, the resulting changeset will be empty
let changeset = graph.apply_update(update);
assert!(changeset.is_empty());

Structs

CanonicalTx

A transaction that is included in the chain, or is still in mempool.

ChangeSet

The ChangeSet represents changes to a TxGraph.

TxAncestors

An iterator that traverses ancestors of a given root transaction.

TxDescendants

An iterator that traverses transaction descendants.

TxGraph

A graph of transactions and spends.

TxNode

A transaction node in the TxGraph.

Enums

CalculateFeeError

Errors returned by TxGraph::calculate_fee.