bdk_chain::tx_graph
Struct ChangeSet

pub struct ChangeSet<A = ()> {
    pub txs: BTreeSet<Arc<Transaction>>,
    pub txouts: BTreeMap<OutPoint, TxOut>,
    pub anchors: BTreeSet<(A, Txid)>,
    pub last_seen: BTreeMap<Txid, u64>,
}

The ChangeSet represents changes to a TxGraph.

Since TxGraph is monotone, the โchangesetโ can only contain transactions to be added and not removed.

Refer to module-level documentation for more.

Fields
txs: BTreeSet<Arc<Transaction>>

Added transactions.

txouts: BTreeMap<OutPoint, TxOut>

Added txouts.

anchors: BTreeSet<(A, Txid)>

Added anchors.

last_seen: BTreeMap<Txid, u64>

Added last-seen unix timestamps of transactions.

Implementations

impl<A> ChangeSet<A>

pub fn txouts(&self) -> impl Iterator<Item = (OutPoint, &TxOut)>

Iterates over all outpoints contained within ChangeSet.

pub fn anchor_heights(&self) -> impl Iterator<Item = u32> + '_
where A: Anchor,

Iterates over the heights of that the new transaction anchors in this changeset.

This is useful if you want to find which heights you need to fetch data about in order to confirm or exclude these anchors.

impl<A: Ord> ChangeSet<A>

pub fn map_anchors<A2: Ord, F>(self, f: F) -> ChangeSet<A2>
where F: FnMut(A) -> A2,

Transform the ChangeSet to have Anchors of another type.

This takes in a closure of signature FnMut(A) -> A2 which is called for each Anchor to transform it.

impl ChangeSet<ConfirmationBlockTime>

pub const SCHEMA_NAME: &'static str = "bdk_txgraph"

Schema name for tx_graph::ChangeSet.

pub const TXS_TABLE_NAME: &'static str = "bdk_txs"

Name of table that stores full transactions and last_seen timestamps.

pub const TXOUTS_TABLE_NAME: &'static str = "bdk_txouts"

Name of table that stores floating txouts.

pub const ANCHORS_TABLE_NAME: &'static str = "bdk_anchors"

Name of table that stores Anchors.

pub fn schema_v0() -> String

Get v0 of sqlite tx_graph::ChangeSet schema

pub fn schema_v1() -> String

Get v1 of sqlite tx_graph::ChangeSet schema

pub fn init_sqlite_tables(db_tx: &Transaction<'_>) -> Result<()>

Initialize sqlite tables.

pub fn from_sqlite(db_tx: &Transaction<'_>) -> Result<Self>

Construct a TxGraph from an sqlite database.

Remember to call Self::init_sqlite_tables beforehand.

pub fn persist_to_sqlite(&self, db_tx: &Transaction<'_>) -> Result<()>

Persist changeset to the sqlite database.

Remember to call Self::init_sqlite_tables beforehand.