Struct bdk_chain::tx_graph::ChangeSet

pub struct ChangeSet<A = ()> {
    pub txs: BTreeSet<Transaction>,
    pub txouts: BTreeMap<OutPoint, TxOut>,
    pub anchors: BTreeSet<(A, Txid)>,
    pub last_seen: BTreeMap<Txid, u64>,
}

The ChangeSet represents changes to a TxGraph.

Since TxGraph is monotone, the “changeset” can only contain transactions to be added and not removed.

Refer to module-level documentation for more.

Fields
txs: BTreeSet<Transaction>

Added transactions.

txouts: BTreeMap<OutPoint, TxOut>

Added txouts.

anchors: BTreeSet<(A, Txid)>

Added anchors.

last_seen: BTreeMap<Txid, u64>

Added last-seen unix timestamps of transactions.

Implementations

impl<A> ChangeSet<A>

pub fn is_empty(&self) -> bool

Returns true if the ChangeSet is empty (no transactions or txouts).

pub fn txouts(&self) -> impl Iterator<Item = (OutPoint, &TxOut)>

Iterates over all outpoints contained within ChangeSet.

pub fn anchor_heights(&self) -> impl Iterator<Item = u32> + '_where
A: Anchor,

Iterates over the heights of that the new transaction anchors in this changeset.

This is useful if you want to find which heights you need to fetch data about in order to confirm or exclude these anchors.

See also: TxGraph::missing_heights

pub fn missing_heights_from<'a>(
&'a self,
local_chain: &'a LocalChain
) -> impl Iterator<Item = u32> + 'awhere
A: Anchor,

Returns an iterator for the anchor_heights in this changeset that are not included in local_chain. This tells you which heights you need to include in local_chain in order for it to conclusively act as a ChainOracle for the transaction anchors this changeset will add.